### 2020-09-17 #2680 @whedon
[PRE REVIEW]: Mangal.jl and EcologicalNetworks.jl: Two complementary packages for analyzing ecological networks in Julia
**Submitting author:** @FrancisBanville (<a href="http://orcid.org/0000-0001-9051-0597">Francis Banville</a>)
**Repository:** <a href="https://github.com/EcoJulia/EcologicalNetworks.jl" target ="_blank">https://github.com/EcoJulia/EcologicalNetworks.jl</a>
**Version:** v0.3.0
