### 2020-09-17 #2680 @kyleniemeyer
Hi @FrancisBanville, thanks for your submission to JOSS. We'll be using this issue to assign the editor and reviewers, and also get any technical/logistical issues out of the way.
