### 2020-09-17 #2680 @kyleniemeyer
@FrancisBanville one issue with your submission is that usually we only review a single software package at a time, though with some exceptions, and I think that your grouping of these two packages together does make sense.

However, you have specified Mangal.jl as the "main" repository for the submission, and your paper is in this repo. Since Mangal.jl would not actually be in the scope of JOSS by itself (as an API wrapper), we would prefer for EcologicalNetworks.jl to be the "main" repository submitted.

We can switch the software repo pointed to internally, but can you also move the paper to the EcologicalNetworks.jl repository? It is fine to live in a branch there, as it is currently in the Mangal.jl repo.

cc @arfon
