+++ paper/paper.bib
@@ -41,14 +41,6 @@
   number = {16}
 }
 
-@misc{Bromberger17,
-  author       = {Seth Bromberger, James Fairbanks, and other contributors},
-  title        = {JuliaGraphs/LightGraphs.jl: an optimized graphs package for the Julia programming language},
-  year         = 2017,
-  doi          = {10.5281/zenodo.889971},
-  url          = {https://doi.org/10.5281/zenodo.889971}
-}
-
 @article{DelmBess19,
   title = {Analysing Ecological Networks of Species Interactions},
   author = {Delmas, Eva and Besson, Mathilde and Brice, Marie Hélène and Burkle, Laura A. and Dalla Riva, Giulio V. and Fortin, Marie Josée and Gravel, Dominique and Guimarães, Paulo R. and Hembry, David H. and Newman, Erica A. and Olesen, Jens M. and Pires, Mathias M. and Yeakel, Justin D. and Poisot, Timothée},
@@ -135,6 +127,24 @@
   eprint = {40835468},
   eprinttype = {jstor},
   number = {4}
+}
+
+@article{MacDBanv20a,
+  title = {Revisiting the {{Links}}-{{Species Scaling Relationship}} in {{Food Webs}}},
+  author = {MacDonald, Arthur Andrew Meahan and Banville, Francis and Poisot, Timothée},
+  date = {2020-07-29},
+  journaltitle = {Patterns},
+  shortjournal = {PATTER},
+  volume = {0},
+  publisher = {{Elsevier}},
+  issn = {2666-3899},
+  doi = {10.1016/j.patter.2020.100079},
+  url = {https://www.cell.com/patterns/abstract/S2666-3899(20)30107-0},
+  urldate = {2020-09-11},
+  file = {C\:\\Users\\franc\\Zotero\\storage\\8Q95IAQF\\MacDonald et al. - 2020 - Revisiting the Links-Species Scaling Relationship .pdf;C\:\\Users\\franc\\Zotero\\storage\\DV9C3TIM\\S2666-3899(20)30107-0.html},
+  keywords = {and tested for one domain/problem,Bayesian methods,constant connectance,DSML 2: Proof-of-Concept: Data science output has been formulated,food webs,implemented,link-species scaling,power law,probabilistic programming},
+  langid = {english},
+  number = {0}
 }
 
 @article{Newm06,
@@ -234,3 +244,5 @@
   langid = {english},
   number = {4}
 }
+
+
