+++ paper/paper.md
@@ -1,21 +1,24 @@
 ---
-title: 'EcologicalNetworks.jl: A Julia package for analyzing networks of species interactions'
+title: 'Mangal.jl and EcologicalNetworks.jl: Two complementary packages for analyzing ecological networks in Julia'
 tags:
   - Julia
   - ecological networks
   - food webs
   - species interactions
-  - graph theory
+  - ecological database
 
 authors:
   - name: Francis Banville^[corresponding author]
     orcid: 0000-0001-9051-0597
     affiliation: "1, 2" # (Multiple affiliations must be quoted)
+  - name: Steve Vissault
+    orcid: 0000-0002-0866-4376
+    affiliation: 2
   - name: Timothée Poisot
     orcid: 0000-0002-0735-5184
     affiliation: 1
 affiliations:
- - name: Deptartment of biology, Université de Montréal, Canada
+ - name: Department of biology, Université de Montréal, Canada
    index: 1
  - name: Department of biology, Université de Sherbrooke, Canada
    index: 2
@@ -25,39 +28,40 @@ bibliography: paper.bib
 
 
 # Summary
-Network ecology is an emerging field of study describing species interactions (e.g. predation, pollination) in a biological community. Ecological networks, in which two species are connected if they can interact, are a mathematical representation of all interactions encountered in a given ecosystem. Anchored in graph theory, the methods developed in network ecology are remarkably rigorous and biologically insightful. Indeed, many ecological and evolutionary processes are driven by species interactions and network structure (i.e. the arrangement of links in ecological networks). The study of ecological networks, from data importation and simulation to data analysis and visualization, require a coherent and efficient set of numerical tools. With its powerful and dynamic type system, the Julia programming language provides a solid computing environment in this regard. `EcologicalNetworks.jl` is a novel Julia package designed to conduct these numerical analysis in an integrated environment for ecological research in Julia.
+Network ecology is an emerging field of study describing species interactions (e.g. predation, pollination) in a biological community. Ecological networks, in which two species are connected if they can interact, are a mathematical representation of all interactions encountered in a given ecosystem. Anchored in graph theory, the methods developed in network ecology are remarkably rigorous and biologically insightful. Indeed, many ecological and evolutionary processes are driven by species interactions and network structure (i.e. the arrangement of links in ecological networks). The study of ecological networks, from data importation and simulation to data analysis and visualization, requires a coherent and efficient set of numerical tools. With its powerful and dynamic type system, the Julia programming language provides a very good computing environment suitable for ecological research. `Mangal.jl` and `EcologicalNetworks.jl` are two novel and complementary Julia packages designed to conduct these numerical analysis efficiently.
 
 # Statement of need
-Analyzing ecological networks is a challenging task. Several measures are commonly used to characterize the structure of ecological networks [@DelmBess19], and their implementation can sometimes be arduous or computationally expensive. The Julia programming language was designed for data science and machine learning, and can thus provide the algorithmic efficiency needed in network ecology. A few packages, such as `LightGraphs.jl` [@Bromberger17], have been written in Julia to analyze any type of graphs. However, the functions provided by these packages are often unsuited for ecological research. When misapplied, they can mislead ecologists in their interpretation of ecological phenomena.
 
[…]
