### 2020-09-17 #2680 @FrancisBanville
Thank you @kyleniemeyer! The paper has been moved to the [EcologicalNetworks.jl repository](https://github.com/EcoJulia/EcologicalNetworks.jl/), in the branch `joss-article`.
