### 2020-09-18 #2680 @kyleniemeyer
@KristinaRiemer absolutely! As editor you don't really need to dive into the software itself—unless you want to!—as long as the reviewers have the appropriate expertise.
