### 2020-09-18 #2680 @KristinaRiemer
I usually do poke around with the code a bit, to learn more about how other folks structure and document their software. I'll be editor for this.
