### 2020-10-05 #2721 @whedon
[REVIEW]: Mangal.jl and EcologicalNetworks.jl: Two complementary packages for analyzing ecological networks in Julia
**Submitting author:** @FrancisBanville (<a href="http://orcid.org/0000-0001-9051-0597">Francis Banville</a>)
**Repository:** <a href="https://github.com/EcoJulia/EcologicalNetworks.jl" target ="_blank">https://github.com/EcoJulia/EcologicalNetworks.jl</a>
**Version:** v0.4.3
