### 2020-10-12 #2721 @arbennett
@FrancisBanville - it seems there's no paper in this repository to review. Is the paper in the Mangal.jl repo?
