### 2020-10-12 #2721 @FrancisBanville
@arbennett - The paper is in the branch `joss-article` of the `EcologicalNetworks.jl` repository.

Sorry for the confusion!
