### 2020-10-13 EcoJulia/EcologicalNetworks.jl#166 @kellykochanski
JOSS: Does authorship match contributions?
The JOSS paper is listed with three authors; however, the github contribution insights show commits from only one of them (tpoisot).

Can you please elaborate on the contributions made to the software by each author? Documentation showing that the contributing author wrote a significant part of the software would be ideal.

As a reminder, JOSS is open to different authorship conventions based on field, but does require that the contributing author has made a "substantial contribution" to the software, and that authors have made more than financial contributions.
