### 2020-10-13 EcoJulia/EcologicalNetworks.jl#167 @kellykochanski
JOSS: Slight clarification of installation instructions
## What to do?

- Add command line instruction for entering package mode in Julia
- If feasible (not required), update requirements to say whether package works with more recent versions of Julia than 1.0

## Why?

To remove ambiguity from language + instructions.
