### 2020-10-13 #2721 @kellykochanski
@KristinaRiemer  I still can't edit the checklist - either I'm missing something simple or it's a bug. I'm definitely logged in, I did accept the invitation (although if I click the link now, it's no longer available - says either the invitation was revoked or I'm in the wrong account?). I've tried both chrome and firefox.
