### 2020-10-13 #2721 @kellykochanski
@FrancisBanville Thanks for submitting a well-written paper and test-covered code repository! I'm ready to approve you on many of the review criteria once I get the checklist working. However, I am also collecting a few questions and issues. These are raised on your repository as issues 166-168:

https://github.com/EcoJulia/EcologicalNetworks.jl/issues/166 - authorship
https://github.com/EcoJulia/EcologicalNetworks.jl/issues/167 - installation instructions
https://github.com/EcoJulia/EcologicalNetworks.jl/issues/168 - paper
