### 2020-10-26 abe079e
add command line instruction for entering package mode in Julia
