### 2020-10-26 6bcaaa6
convert use cases to present tense
