+++ paper/paper.md
@@ -61,7 +61,7 @@ Table: Description of all custom types in `Mangal.jl`. Related custom types are 
 
 Understanding how the number of interactions scales with the number of species is fundamental in ecology. This relationship has recently been revisited using all food webs archived on `mangal.io` [@MacDBanv20a], making it, to the best of our knowledge, the most extensive study of such relationship so far. The block of code below read relevant metadata from `mangal.io` to conduct this analysis on all types of ecological networks.
 
-We first retrieved all networks archived on the database, which returned objects of type `MangalNetwork`. We then counted the number of species $S$ and the total number of interactions $L$ in each network, as well as their numbers of interactions of predation, of herbivory, of mutualism, and of parasitism. These information were stored in a data frame along with the ID numbers of the networks.
+We first retrieve all networks archived on the database, which returns objects of type `MangalNetwork`. We then count the number of species $S$ and the total number of interactions $L$ in each network, as well as their numbers of interactions of predation, of herbivory, of mutualism, and of parasitism. We store these information in a data frame along with the ID numbers of the networks.
 
 ```julia
 
@@ -101,9 +101,9 @@ global cursor = 1
 end
 ```
 
-The association between the number of species $S$ and the total number of links $L$ is plotted in \autoref{fig:LS}. We classified all networks according to their most frequent type of links, and considered interactions of predation and of herbivory as food-web interactions. Networks in the "other types" category include interactions between competitors, symbiotes, scavengers, and detritivores, among others. Very small networks (i.e. with less than 5 interactions) were discarded.  
+The association between the number of species $S$ and the total number of links $L$ is plotted in \autoref{fig:LS}. We classify all networks according to their most frequent type of links, and consider interactions of predation and of herbivory as food-web interactions. Networks in the "other types" category include interactions between competitors, symbiotes, scavengers, and detritivores, among others. Very small networks (i.e. with less than 5 interactions) are discarded.  
 
-![Association between the number of species (nodes) and the number of links (edges) in ecological networks archived on `mangal.io`. Networks with less than 5 interactions were discarded. Different types of interactions can be listed within the same network. We classified all networks according to their most frequent type of interactions. The code to reproduce the figure is included in this paper's repository.\label{fig:LS}](fig/LS.png)
+![Association between the number of species (nodes) and the number of links (edges) in ecological networks archived on `mangal.io`. Networks with less than 5 interactions are discarded. Different types of interactions can be listed within the same network. We classify all networks according to their most frequent type of interactions. The code to reproduce the figure is included in this paper's repository.\label{fig:LS}](fig/LS.png)
 
 
 # EcologicalNetworks.jl
@@ -117,7 +117,7 @@ The package [`EcologicalNetworksPlots.jl`](https://github.com/EcoJulia/Ecologica
 
[…]
