### 2020-10-26 fbb0a67
change a diversity of measures to diverse measures
