+++ paper/paper.md
@@ -32,7 +32,7 @@ Network ecology is an emerging field of study describing species interactions (e
 
 # Statement of need
 
-Analyzing ecological networks is a challenging task. First, data retrieval and manipulation can be difficult, as authors typically store their data in different repositories and in different formats. Further, a diversity of measures is commonly used to characterize the structure of ecological networks [@DelmBess19], and their implementation can sometimes be arduous or computationally expensive. The Julia programming language was designed for data science and machine learning, and can thus provide the algorithmic efficiency needed in ecology. [EcoJulia](https://ecojulia.github.io/), an integrated environment for the conduction of ecological research in Julia, contains a set of packages that can be used for a variety of ecological topics, from spatial analysis to phylogenetics.
+Analyzing ecological networks is a challenging task. First, data retrieval and manipulation can be difficult, as authors typically store their data in different repositories and in different formats. Further, diverse measures are commonly used to characterize the structure of ecological networks [@DelmBess19], and their implementation can sometimes be arduous or computationally expensive. The Julia programming language was designed for data science and machine learning, and can thus provide the algorithmic efficiency needed in ecology. [EcoJulia](https://ecojulia.github.io/), an integrated environment for the conduction of ecological research in Julia, contains a set of packages that can be used for a variety of ecological topics, from spatial analysis to phylogenetics.
 
 Together, two interrelated packages in EcoJulia provide a useful methodological framework for the analysis of ecological networks: [`Mangal.jl`](https://github.com/EcoJulia/Mangal.jl) and [`EcologicalNetworks.jl`](https://github.com/EcoJulia/EcologicalNetworks.jl). The [Mangal project](https://mangal.io/#/) was set up to meet the need for an extensive ecological interaction database with an easy data retrieval system. In Julia, `Mangal.jl` is used to read data from `mangal.io`. Once imported, these networks can then be analyzed using `EcologicalNetworks.jl`, which could also be used to simulate ecological networks under different statistical and ecological models. Here we provide an overview of the functionalities of these two packages, and illustrate how they can be used to answer fundamental ecological questions.
 
