### 2020-10-26 4b33096
explain why Julia's typing system is good for ecological network analysis
