+++ paper/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 
 # Summary
-Network ecology is an emerging field of study describing species interactions (e.g. predation, pollination) in a biological community. Ecological networks, in which two species are connected if they can interact, are a mathematical representation of all interactions encountered in a given ecosystem. Anchored in graph theory, the methods developed in network ecology are remarkably rigorous and biologically insightful. Indeed, many ecological and evolutionary processes are driven by species interactions and network structure (i.e. the arrangement of links in ecological networks). The study of ecological networks, from data importation and simulation to data analysis and visualization, requires a coherent and efficient set of numerical tools. With its powerful and dynamic type system, the Julia programming language provides a very good computing environment suitable for ecological research. `Mangal.jl` and `EcologicalNetworks.jl` are two novel and complementary Julia packages designed to conduct these numerical analysis efficiently.
+Network ecology is an emerging field of study describing species interactions (e.g. predation, pollination) in a biological community. Ecological networks, in which two species are connected if they can interact, are a mathematical representation of all interactions encountered in a given ecosystem. Anchored in graph theory, the methods developed in network ecology are remarkably rigorous and biologically insightful. Indeed, many ecological and evolutionary processes are driven by species interactions and network structure (i.e. the arrangement of links in ecological networks). The study of ecological networks, from data importation and simulation to data analysis and visualization, requires a coherent and efficient set of numerical tools. With its powerful and dynamic [type system](https://docs.julialang.org/en/v1/manual/types/), the Julia programming language provides a very good and fast computing environment suitable for ecological research. Julia's typing system notably ensures that appropriate methods are used when a function is applied with different types of ecological data, which therefore minimizes the risk of committing errors of analysis and interpretation in network ecology. `Mangal.jl` and `EcologicalNetworks.jl` are two novel and complementary Julia packages designed to conduct research in network ecology efficiently.
 
 # Statement of need
 
