### 2020-10-26 fc5dd51
describe the uses of the package more specifically in the summary
