+++ paper/paper.bib
@@ -129,6 +129,23 @@
   number = {4}
 }
 
+@article{Guim20,
+  title = {The {{Structure}} of {{Ecological Networks Across Levels}} of {{Organization}}},
+  author = {Guimarães, Paulo R.},
+  date = {2020-11-02},
+  journaltitle = {Annual Review of Ecology, Evolution, and Systematics},
+  shortjournal = {Annu. Rev. Ecol. Evol. Syst.},
+  volume = {51},
+  pages = {annurev-ecolsys-012220-120819},
+  issn = {1543-592X, 1545-2069},
+  doi = {10.1146/annurev-ecolsys-012220-120819},
+  url = {https://www.annualreviews.org/doi/10.1146/annurev-ecolsys-012220-120819},
+  urldate = {2020-09-02},
+  abstract = {Interactions connect the units of ecological systems, forming networks. Individual-based networks characterize variation in niches among individuals within populations. These individual-based networks merge with each other, forming species-based networks and food webs that describe the architecture of ecological communities. Networks at broader spatiotemporal scales portray the structure of ecological interactions across landscapes and over macroevolutionary time. Here, I review the patterns observed in ecological networks across multiple levels of biological organization. A fundamental challenge is to understand the amount of interdependence as we move from individual-based networks to species-based networks and beyond. Despite the uneven distribution of studies, regularities in network structure emerge across scales due to the fundamental architectural patterns shared by complex networks and the interplay between traits and numerical effects. I illustrate the integration of these organizational scales by exploring the consequences of the emergence of highly connected species for network structures across scales.             Expected final online publication date for the Annual Review of Ecology, Evolution, and Systematics, Volume 51 is November 2, 2020. Please see http://www.annualreviews.org/page/journal/pubdates for revised estimates.},
+  langid = {english},
+  number = {1}
+}
+
 @article{MacDBanv20a,
   title = {Revisiting the {{Links}}-{{Species Scaling Relationship}} in {{Food Webs}}},
   author = {MacDonald, Arthur Andrew Meahan and Banville, Francis and Poisot, Timothée},
