+++ paper/paper.bib
@@ -52,6 +52,17 @@
   keywords = {biogeography,community ecology,ecological networks,graph theory,interactions}
 }
 
+@article{DormGrub08,
+  title = {Introducing the Bipartite {{Package}}: {{Analysing Ecological Networks}}},
+  shorttitle = {Introducing the Bipartite {{Package}}},
+  author = {Dormann, Carsten and Gruber, Bernd and Fründ, Jochen},
+  date = {2008-01-01},
+  journaltitle = {R News},
+  shortjournal = {R News},
+  volume = {8},
+  file = {C\:\\Users\\franc\\Zotero\\storage\\K6BKFD2D\\Dormann et al. - 2008 - Introducing the bipartite Package Analysing Ecolo.pdf}
+}
+
 @article{DunnWill02c,
   title = {Network Structure and Biodiversity Loss in Food Webs: Robustness Increases with Connectance},
   shorttitle = {Network Structure and Biodiversity Loss in Food Webs},
@@ -92,6 +103,26 @@
   file = {C\:\\Users\\franc\\Zotero\\storage\\FRKZUBPR\\Dunne et al. - 2002 - Food-web structure and network theory The role of.pdf;C\:\\Users\\franc\\Zotero\\storage\\M5W7ZZF8\\12917.html},
   langid = {english},
   number = {20}
+}
+
+@article{FlorPois16,
+  title = {{{BiMat}}: A {{MATLAB}} Package to Facilitate the Analysis of Bipartite Networks},
+  shorttitle = {{{BiMat}}},
+  author = {Flores, Cesar O. and Poisot, Timothée and Valverde, Sergi and Weitz, Joshua S.},
+  date = {2016},
+  journaltitle = {Methods in Ecology and Evolution},
+  volume = {7},
+  pages = {127--132},
+  issn = {2041-210X},
+  doi = {10.1111/2041-210X.12458},
+  url = {https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/2041-210X.12458},
+  urldate = {2020-10-27},
+  abstract = {Bipartite networks are ubiquitous in community ecology, including examples of facilitative interactions, such as plant-pollinator networks, and antagonistic interactions, such as virus–host infection networks. Statistical network analysis is increasingly used to identify emergent, nonrandom patterns of interaction and the effect of interaction patterns on ecological and evolutionary dynamics. Two recurring patterns are that of modularity and nestedness. Modularity is a feature of networks in which there are densely interacting subgroups. Nestedness is a feature of networks in which the interactions form ordered subsets. Here we describe BiMat, an open-source MATLAB package for the study of the structure of bipartite ecological networks. Unlike alternative tools, BiMat enables both multiscale analysis of the structure of a bipartite ecological network – spanning global (i.e. entire network) to local (i.e. module-level) scales – and meta-analyses of many bipartite networks simultaneously. In common with other tools, BiMat calculates the degree to which a bipartite network is modular and/or nested, the statistical significance of these patterns, and enables visualization of latent structures in the network. BiMat is available as an open-source MATLAB package, with a quick-start guide and worked examples at: http://bimat.github.io/.},
+  annotation = {\_eprint: https://besjournals.onlinelibrary.wiley.com/doi/pdf/10.1111/2041-210X.12458},
+  file = {C\:\\Users\\franc\\Zotero\\storage\\PHAXW54C\\Flores et al. - 2016 - BiMat a MATLAB package to facilitate the analysis.pdf;C\:\\Users\\franc\\Zotero\\storage\\2IXGAMLE\\2041-210X.html},
[…]
