### 2020-10-28 EcoJulia/EcologicalNetworks.jl#166 @tpoisot
Thank you for pointing this out - @francisbanville is starting a *Ph.D.* on ecological networks, and his project implies that he will take over most of the development of `EcologicalNetworks.jl`. Writing this article as a way to initiate this process.

It must also be noted that @francisbanville has been *very* active in outreach on the package, and notably gave a talk at the latest JuliaCon meeting, which formed the basis for the content of the article. This is not a "code" contribution, but I think that this is a very important one for the adoption of the tool.
