### 2020-10-28 EcoJulia/EcologicalNetworks.jl#166 @FrancisBanville
Thanks @tpoisot.

I also want to specify that @SteveViss has been the main contributor to @mangal-wg. His contribution to Mangal was substantial, and this is why he is an author of this article.
