### 2020-10-28 EcoJulia/EcologicalNetworks.jl#166 @SteveViss
Thanks @FrancisBanville for flagging me out in this issue. @kellykochanski, I redesigned the mangal v2 backend (database + REST API) which feeds the client `mangal.jl` and allows the `EcologicalNetworks.jl` to exist. Let me know if you need further details on my contribution.
