### 2020-11-02 7bfe3fd
specify supported versions of julia in manuscript
