+++ paper/paper.md
@@ -34,7 +34,9 @@ Network ecology is an emerging field of study describing species interactions (e
 
 Ecological networks are a gold mine of biological information. From defining species' roles within the context of their community to describing the functioning of ecosystems, the analysis of ecological networks can help us predict how species or entire ecosystems will respond to environmental changes [@DelmBess19; @Guim20]. Analyzing ecological networks is however a very challenging task. First, data retrieval and manipulation can be difficult, as authors typically store their data in different repositories and in different formats. Further, diverse measures are commonly used to characterize the structure of ecological networks [@DelmBess19], and their implementation can sometimes be arduous or computationally expensive. The Julia programming language was designed for data science and machine learning, and can thus provide the algorithmic efficiency needed in ecology. [EcoJulia](https://ecojulia.github.io/), an integrated environment for the conduction of ecological research in Julia, contains a set of packages that can be used for a variety of ecological topics, from spatial analysis to phylogenetics.
 
-Together, two interrelated packages in EcoJulia provide a useful methodological framework for the analysis of ecological networks: [`Mangal.jl`](https://github.com/EcoJulia/Mangal.jl) and [`EcologicalNetworks.jl`](https://github.com/EcoJulia/EcologicalNetworks.jl). The [Mangal project](https://mangal.io/#/) was set up to meet the need for an extensive ecological interaction database with an easy data retrieval system. In Julia, `Mangal.jl` is used to read data from `mangal.io`. Once imported, these networks can then be analyzed using `EcologicalNetworks.jl`, which could also be used to simulate ecological networks under different statistical and ecological models. `Mangal.jl` and `EcologicalNetworks.jl` are designed to work with the most recent versions of Julia (Julia 1.3+). Here we provide an overview of the functionalities of these two packages, and illustrate how they can be used to answer fundamental ecological questions.
+Together, two interrelated packages in EcoJulia provide a useful methodological framework for the analysis of ecological networks: [`Mangal.jl`](https://github.com/EcoJulia/Mangal.jl) and [`EcologicalNetworks.jl`](https://github.com/EcoJulia/EcologicalNetworks.jl). The [Mangal project](https://mangal.io/#/) was set up to meet the need for an extensive ecological interaction database with an easy data retrieval system. In Julia, `Mangal.jl` is used to read data from `mangal.io`. Once imported, these networks can then be analyzed using `EcologicalNetworks.jl`, which could also be used to simulate ecological networks under different statistical and ecological models. Our suggested methodological framework is described in more detail in \autoref{fig:workflow}. `Mangal.jl` and `EcologicalNetworks.jl` are designed to work with the most recent versions of Julia (Julia 1.3+). Here we provide an overview of the functionalities of these two packages, and illustrate how they can be used to answer fundamental ecological questions.
+
+![**Proposed methodological framework for the analysis of ecological networks in Julia.** **Inputs**: More than 1300 ecological networks of various types are archived on the `mangal.io` database. These networks are read in Julia using `Mangal.jl`. **Outputs from `Mangal.jl`**: Objects (e.g. species, interactions, metadata) can be counted directly using `Mangal.jl`. **Outputs from `EcologicalNetworks.jl`**: All types of networks, including probabilistic networks, can be further analyzed using `EcologicalNetworks.jl`. The main measures and analysis supported by this package include species richness, connectance, linkage density, species' degree, centrality measures, trophic level analysis, shortest paths, nestedness, motif counting, modularity, β-diversity measures, and information theoretic analysis. Empirical networks can then be compared to a random network distribution simulated using null models, constrained network permutations, and structural random models. **Backend methods**: These measures and analysis methods are described in further depth in @DelmBess19, @PoisCirt16, and @PoisBeli19a. \label{fig:workflow}](fig/workflow.png)
 
 ## State of the field
 
