### 2020-11-03 EcoJulia/EcologicalNetworks.jl#166 @KristinaRiemer
All of these authorship contributions are sufficient according to JOSS standards, thanks to everyone for clarifying their roles!
