### 2020-11-04 b5f5aa0
update workflow figure
