### 2020-11-04 EcoJulia/EcologicalNetworks.jl#167 @FrancisBanville
Thanks for your comment @kellykochanski!

Command line instructions for entering package mode in Julia have been added to the `Mangal.jl` and `EcologicalNetworks.jl` repositories. I also specified in the manuscript which versions of Julia are supported.
