### 2020-11-04 EcoJulia/EcologicalNetworks.jl#168 @FrancisBanville
Thank you so much for your valuable feedbacks @kellykochanski.

The modified manuscript is in the branch [`joss-article-review1`](https://github.com/EcoJulia/EcologicalNetworks.jl/tree/joss-article-review1). We hope that we have adequately addressed your comments.

### Summary
We added a sentence in the summary describing in greater detail the uses of the software.

We chose to keep the statement that Julia's typing system is good for ecological research, and we explained more specifically why.

### Statement of need
We added a brief explanation at the beginning of this section about the ultimate impacts of analyzing ecological networks.

### State of the field
We added a section on the state of the field, in which we listed the main packages available for modeling ecological networks, in Julia and in other programming languages. We explained how our software is different and what are its added values.

### Quality of writing
We added a figure on our proposed workflow, in which we explained the inputs, outputs, and analysis methods more clearly.

We also wrote our use cases in the present tense.
