### 2020-11-04 #2721 @FrancisBanville
> @FrancisBanville Thanks for submitting a well-written paper and test-covered code repository! I'm ready to approve you on many of the review criteria once I get the checklist working [edit: now done]. However, I am also collecting a few questions and issues. These are raised on your repository as issues 166-168:
>
> * [ ]  [EcoJulia/EcologicalNetworks.jl#166](https://github.com/EcoJulia/EcologicalNetworks.jl/issues/166) - contribution and authorship
> * [ ]  [EcoJulia/EcologicalNetworks.jl#167](https://github.com/EcoJulia/EcologicalNetworks.jl/issues/167) - installation instructions
> * [ ]  [EcoJulia/EcologicalNetworks.jl#168](https://github.com/EcoJulia/EcologicalNetworks.jl/issues/168) - software paper
>
> I'm aiming to raise small issues, each relating to one or more specific JOSS review criteria, to streamline the editing and approval process.

Thank you for your helpful review @kellykochanski, and sorry for the late reply!

We have made some changes to the manuscript (see branch [`joss-article-review1`](https://github.com/EcoJulia/EcologicalNetworks.jl/tree/joss-article-review1)) and commented on your issues. Please do not hesitate to contact us if you have any questions or if you need further detail!
