### 2020-11-09 EcoJulia/EcologicalNetworks.jl#170 @arbennett
JOSS: EcoJulia link doesn't work.
In the JOSS paper text there is a link to EcoJulia given as https://ecojulia.github.io/ - however, I am getting "Site not found".
