### 2020-11-09 EcoJulia/EcologicalNetworks.jl#171 @arbennett
JOSS: Include code example in docs
One of the review criteria for the JOSS paper in the documentation section is:
> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)

Currently the docs don't have a concrete example that would help new users get up and running. The code examples given in the paper itself, however, are really nice and could be an easy inclusion into the docs. I think that it would be really helpful to have this example in the docs, or even just a link to the script in the paper directory.
