### 2020-11-09 #2721 @arbennett
@KristinaRiemer - sorry for the slow review. I think my review is now complete.

@FrancisBanville - thanks for this submission. The paper is well written and covers all of the required bases (I just re-looked at it in the revised branch, and also really like the added diagram). Your example use cases are very nice and prove this is robust and mature software. The code quality, from my quick, shallow dive, is really good and I like how you've described your type system hierarchy, which definitely seems to be a boon for Julia. The documentation that you've written up is also quite clean and slick, and provides both technical as well as intuitive aspects of your software design. I only have two minor considerations that I would like to see addressed before accepting this for submission. They are:

- [ ] [ JOSS: EcoJulia link doesn't work. #170 ](https://github.com/EcoJulia/EcologicalNetworks.jl/issues/170)
- [ ] [ JOSS: Include code example in docs #171 ](https://github.com/EcoJulia/EcologicalNetworks.jl/issues/171)

I also was unclear on the contribution and authorship question brought up by @kellykochanski - perhaps @KristinaRiemer or @danielskatz could chime in on this? To me, the response in [#166](https://github.com/EcoJulia/EcologicalNetworks.jl/issues/166) seems reasonable and I'm happy with it.
