### 2020-11-10 EcoJulia/EcologicalNetworks.jl#166 @kellykochanski
This looks good. Thank you for clarifying!
