### 2020-11-10 EcoJulia/EcologicalNetworks.jl#171 @tpoisot
@FrancisBanville I think it can be a good idea to have a "Use-cases" section like in the Mangal doc, with some examples that are specific to this package. Thanks @arbennett for flagging this!
