### 2020-11-10 EcoJulia/EcologicalNetworks.jl#170 @tpoisot
Good catch - the new URL is `https://ecojulia.org/` -- @FrancisBanville this is a trivial fix.
