+++ paper/paper.bib
@@ -322,20 +322,4 @@
   number = {4}
 }
 
-@misc{Bromberger17,
-  author       = {Seth Bromberger, James Fairbanks, and other contributors},
-  title        = {JuliaGraphs/LightGraphs.jl: an optimized graphs package for the Julia programming language},
-  year         = 2017,
-  doi          = {10.5281/zenodo.889971},
-  url          = {https://doi.org/10.5281/zenodo.889971}
-}
 
-@Article{CsarNepu06,
-   title = {The igraph software package for complex network research},
-   author = {Gabor Csardi and Tamas Nepusz},
-   journal = {InterJournal},
-   volume = {Complex Systems},
-   pages = {1695},
-   year = {2006},
-   url = {https://igraph.org},
- }
