### 2020-11-10 EcoJulia/EcologicalNetworks.jl#170 @FrancisBanville
@arbennett Thank you for catching that! I just fixed it.
