### 2020-11-10 6345d25
example use case - integration with mangal
