+++ docs/src/examples/integration_mangal.md
@@ -1,0 +1,178 @@
+# Integration with Mangal.jl
+
+Here we provide an example use case to get you started with `EcologicalNetworks.jl`. We show how to analyse meaningful network properties using all networks archived on the [`mangal.io`](https://mangal.io/#/) online database. We will create two plots: (1) the association between the number of species and the number of links in different types of networks, and (2) the association between species richness, connectance, nestedness, and modularity in food webs.
+
+To conduct these analyses, we need to upload the following packages:
+
+```julia
+using EcologicalNetworks
+using Mangal
+using DataFrames
+using Plots
+```
+
+We first retrieve relevant metadata for all networks archived on `mangal.io`. We count the number of species $S$ and the total number of interactions $L$ in each network, as well as their numbers of interactions of predation, of herbivory, of mutualism, and of parasitism. We store these information in a data frame along with the networks' ID numbers.
+
+```julia
+number_of_networks = count(MangalNetwork)
+count_per_page = 100
+number_of_pages = convert(Int, ceil(number_of_networks/count_per_page))
+
+mangal_networks = DataFrame(fill(Int64, 7),
+                 [:id, :S, :L, :pred, :herb, :mutu, :para],
+                 number_of_networks)
+
+global cursor = 1
+@progress "Paging networks" for page in 1:number_of_pages
+    global cursor
+    networks_in_page = Mangal.networks("count" => count_per_page, "page" => page-1)
+    @progress "Counting items" for current_network in networks_in_page
+        S = count(MangalNode, current_network)
+        L = count(MangalInteraction, current_network)
+        pred = count(MangalInteraction, current_network, "type" => "predation")
+        herb = count(MangalInteraction, current_network, "type" => "herbivory")
+        mutu = count(MangalInteraction, current_network, "type" => "mutualism")
+        para = count(MangalInteraction, current_network, "type" => "parasitism")
+        mangal_networks[cursor,:] .= (current_network.id, S, L, pred, herb, mutu, para)
+        cursor = cursor + 1
+    end
+end
+```
+
+We can classify all networks according to their most frequent type of interactions. Note that trophic interactions in food webs comprise interactions of predation and of herbivory. Note also that we discard small networks having less than 5 interactions for reliability reasons.
+
+```julia
+# number of trophic interactions
+mangal_networks.foodweb = mangal_networks.pred .+ mangal_networks.herb
+
+# number of other types of interactions
+mangal_networks.other = mangal_networks.L .- mangal_networks.foodweb .- mangal_networks.mutu .- mangal_networks.para
+
+# discard networks with less than 5 links
+mangal_networks2 = mangal_networks[mangal_networks.L .> 4, :]
+
+# classify networks
+number_interactions_max_type = maximum.(eachrow(mangal_networks2[:,[:mutu, :para,:foodweb,:other]]))
+
+foodwebs = mangal_networks2[mangal_networks2[:foodweb] .== number_interactions_max_type ,:]
+parasitism_webs = mangal_networks2[mangal_networks2[:para] .== number_interactions_max_type ,:]
+mutualism_webs = mangal_networks2[mangal_networks2[:mutu] .== number_interactions_max_type ,:]
+other_webs = mangal_networks2[mangal_networks2[:other] .== number_interactions_max_type ,:]
+```
+
[…]
