### 2020-11-10 6f69c17
:memo: update name of the use case section in the doc
