+++ test/modularity/roles.jl
@@ -13,23 +13,23 @@ using EcologicalNetworks
         false false false false false false
         ]
     U = UnipartiteNetwork(A)
-    _, L = brim(lp(U)...)
-    n = functional_cartography(N, L)
+    L = Dict("s1"=>1, "s2"=>1, "s3"=>1, "s4"=>2, "s5"=>2, "s6"=>2)
+    n = functional_cartography(U, L)
 
     # Values of z
-    @test n[1,1] ≈ -0.57 atol = 0.02
-    @test n[2,1] ≈  1.15 atol = 0.02
-    @test n[3,1] ≈ -0.57 atol = 0.02
-    @test n[4,1] ≈ -0.57 atol = 0.02
-    @test n[5,1] ≈  1.15 atol = 0.02
-    @test n[6,1] ≈ -0.57 atol = 0.02
+    @test n["s1"][1] ≈ -0.57 atol = 0.02
+    @test n["s2"][1] ≈  1.15 atol = 0.02
+    @test n["s3"][1] ≈ -0.57 atol = 0.02
+    @test n["s4"][1] ≈ -0.57 atol = 0.02
+    @test n["s5"][1] ≈  1.15 atol = 0.02
+    @test n["s6"][1] ≈ -0.57 atol = 0.02
 
     # Values of c
-    @test n[1,2] ≈ 0.00 atol = 0.02
-    @test n[2,2] ≈ 0.45 atol = 0.02
-    @test n[3,2] ≈ 0.00 atol = 0.02
-    @test n[4,2] ≈ 0.50 atol = 0.02
-    @test n[5,2] ≈ 0.00 atol = 0.02
-    @test n[6,2] ≈ 0.00 atol = 0.02
+    @test n["s1"][2] ≈ 0.00 atol = 0.02
+    @test n["s2"][2] ≈ 0.45 atol = 0.02
+    @test n["s3"][2] ≈ 0.00 atol = 0.02
+    @test n["s4"][2] ≈ 0.50 atol = 0.02
+    @test n["s5"][2] ≈ 0.00 atol = 0.02
+    @test n["s6"][2] ≈ 0.00 atol = 0.02
 
 end
