### 2020-11-11 b5e084d
:bug: Fixes #162 (NaN std for functional roles)
