### 2020-11-11 9cd857a
:memo: update example use case
