+++ docs/src/examples/integration_mangal.md
@@ -1,25 +1,25 @@
 # Integration with Mangal.jl
 
-Here we provide an example use case to get you started with `EcologicalNetworks.jl`. We show how to analyse meaningful network properties using all networks archived on the [`mangal.io`](https://mangal.io/#/) online database. We will create two plots: (1) the association between the number of species and the number of links in different types of networks, and (2) the association between species richness, connectance, nestedness, and modularity in food webs.
+In this example, we will show how `EcologicalNetworks.jl` can be integrated with [`Mangal.jl`](https://github.com/EcoJulia/Mangal.jl) to analyse many ecological networks. Specifically, we will show how to analyse the association between meaningful network properties (i.e. species richness, connectance, nestedness, and modularity) using all food webs archived on the [`mangal.io`](https://mangal.io/#/) online database.
 
-To conduct these analyses, we need to upload the following packages:
+To conduct this analysis, we need to upload the following packages:
 
-```julia
+```@example mangal
 using EcologicalNetworks
 using Mangal
 using DataFrames
 using Plots
 ```
 
-We first retrieve relevant metadata for all networks archived on `mangal.io`. We count the number of species $S$ and the total number of interactions $L$ in each network, as well as their numbers of interactions of predation, of herbivory, of mutualism, and of parasitism. We store these information in a data frame along with the networks' ID numbers.
+We first retrieve relevant metadata for all networks archived on `mangal.io` using the `Mangal.jl` package. We count the number of species $S$ and the total number of interactions $L$ in each network, as well as their number of trophic interactions (predation and herbivory). We store these information in a data frame along with the networks' ID numbers.
 
-```julia
+```@example mangal
 number_of_networks = count(MangalNetwork)
 count_per_page = 100
 number_of_pages = convert(Int, ceil(number_of_networks/count_per_page))
 
-mangal_networks = DataFrame(fill(Int64, 7),
-                 [:id, :S, :L, :pred, :herb, :mutu, :para],
+mangal_networks = DataFrame(fill(Int64, 5),
+                 [:id, :S, :L, :pred, :herb],
                  number_of_networks)
 
 global cursor = 1
@@ -31,75 +31,24 @@ global cursor = 1
         L = count(MangalInteraction, current_network)
         pred = count(MangalInteraction, current_network, "type" => "predation")
         herb = count(MangalInteraction, current_network, "type" => "herbivory")
-        mutu = count(MangalInteraction, current_network, "type" => "mutualism")
-        para = count(MangalInteraction, current_network, "type" => "parasitism")
-        mangal_networks[cursor,:] .= (current_network.id, S, L, pred, herb, mutu, para)
+        mangal_networks[cursor,:] .= (current_network.id, S, L, pred, herb)
         cursor = cursor + 1
     end
 end
 ```
 
-We can classify all networks according to their most frequent type of interactions. Note that trophic interactions in food webs comprise interactions of predation and of herbivory. Note also that we discard small networks having less than 5 interactions for reliability reasons.
+We now have all the information we need to identify all food webs archived on `mangal.io`. Here we consider as food webs any ecological networks mainly composed of trophic interactions.
 
-```julia
-# number of trophic interactions
-mangal_networks.foodweb = mangal_networks.pred .+ mangal_networks.herb
-
[…]
