### 2020-11-12 e4b8ba2
updates to the style
