+++ paper/paper.bib
@@ -41,6 +41,28 @@
   number = {16}
 }
 
+@software{BromOthe17,
+  title = {Juliagraphs/{{Lightgraphs}}.{{Jl}}},
+  author = {Bromberger, Seth and Other Contributors},
+  date = {2017-09-16},
+  doi = {10.5281/ZENODO.889971},
+  url = {https://zenodo.org/record/889971},
+  urldate = {2020-11-12},
+  abstract = {LightGraphs: An Optimized Graphs Package for the Julia Programming Language},
+  organization = {{Zenodo}}
+}
+
+@article{CsarNepu05,
+  title = {The {{Igraph Software Package}} for {{Complex Network Research}}},
+  author = {Csardi, Gabor and Nepusz, Tamas},
+  date = {2005-11-30},
+  journaltitle = {InterJournal},
+  shortjournal = {InterJournal},
+  volume = {Complex Systems},
+  pages = {1695},
+  file = {C\:\\Users\\franc\\Zotero\\storage\\UEXUPG5Z\\Csardi et Nepusz - 2005 - The Igraph Software Package for Complex Network Re.pdf}
+}
+
 @article{DelmBess19,
   title = {Analysing Ecological Networks of Species Interactions},
   author = {Delmas, Eva and Besson, Mathilde and Brice, Marie Hélène and Burkle, Laura A. and Dalla Riva, Giulio V. and Fortin, Marie Josée and Gravel, Dominique and Guimarães, Paulo R. and Hembry, David H. and Newman, Erica A. and Olesen, Jens M. and Pires, Mathias M. and Yeakel, Justin D. and Poisot, Timothée},
