+++ paper/paper.md
@@ -40,9 +40,9 @@ Together, two interrelated packages in EcoJulia provide a useful methodological 
 
 ## State of the field
 
-The main software packages available for analyzing ecological networks include [`igraph`](https://igraph.org/redirect.html) [@CsarNepu06; Python, R, and C], [`networkx`](https://networkx.org/documentation/stable/index.html) [@HagbSchu08; Python], [`bipartite`](https://cran.r-project.org/web/packages/bipartite/index.html) [@DormGrub08; R], [`cheddar`](https://cran.r-project.org/web/packages/cheddar/index.html) [@HudsEmer13; R], and [`BiMat`](https://bimat.github.io/) [@FlorPois16; Octave]. To the best of our knowledge, `EcologicalNetworks.jl` is the only software that supports all types of ecological networks, and the only one that supports probabilistic ecological networks.
+The main software packages available for analyzing ecological networks include [`igraph`](https://igraph.org/redirect.html) [@CsarNepu05; Python, R, and C], [`networkx`](https://networkx.org/documentation/stable/index.html) [@HagbSchu08; Python], [`bipartite`](https://cran.r-project.org/web/packages/bipartite/index.html) [@DormGrub08; R], [`cheddar`](https://cran.r-project.org/web/packages/cheddar/index.html) [@HudsEmer13; R], and [`BiMat`](https://bimat.github.io/) [@FlorPois16; Octave]. To the best of our knowledge, `EcologicalNetworks.jl` is the only software that supports all types of ecological networks, and the only one that supports probabilistic ecological networks.
 
-In Julia, the [JuliaGraphs](https://github.com/JuliaGraphs) ecosystem contains a set of performant packages for modeling and analyzing different types of graphs and networks from many disciplines. The main JuliaGraphs package is [`LightGraphs.jl`](https://juliagraphs.org/LightGraphs.jl/stable/) [@Bromberger17], which implements functions for the analysis of simple undirected and directed graphs.
+In Julia, the [JuliaGraphs](https://github.com/JuliaGraphs) ecosystem contains a set of performant packages for modeling and analyzing different types of graphs and networks from many disciplines. The main JuliaGraphs package is [`LightGraphs.jl`](https://juliagraphs.org/LightGraphs.jl/stable/) [@BromOthe17], which implements functions for the analysis of simple undirected and directed graphs.
 
 The added-value of using `EcologicalNetworks.jl` to analyze ecological networks, instead of a more general-purpose package like `LightGraphs.jl`, is two-fold: (1) its integration with other EcoJulia packages, including `Mangal.jl`, and with ecological databases, greatly simplifies the conduction of comprehensive ecological research in Julia, and (2) most functions of a general-purpose graph analysis package may not be suitable to adequately address many ecological questions. Indeed, functions provided by `EcologicalNetworks.jl` are based on state-of-the-art practices in network ecology [@PoisCirt16; @DelmBess19], whereas general-purpose functions might be misleading when it comes to interpreting them ecologically.
 
