### 2020-11-12 79dec8f
add mangal to doc dependencies
