### 2020-11-12 206aa21
add DataFrames and Plots to doc dependencies
