+++ docs/make.jl
@@ -31,6 +31,7 @@ makedocs(
                                         "Structural models" => "random/structure.md"
                                        ],
                   "Examples" => [
+                                "Modularity" => "examples/modularity.md",
                                 "Integration with Mangal" => "examples/integration_mangal.md",
                                 "Network beta-diversity" => "examples/beta-diversity.md"
                                  ]
