+++ docs/src/examples/modularity.md
@@ -1,20 +1,23 @@
-## Modularity
+# Modularity
 
-In this example, we will show how the modular structure of an ecological
-network can be optimized. Finding the optimal modular structure can be a
-time-consuming process, as it relies on heuristic which are not guaranteed to
-converge to the global maximum. There is no elegant alternative to trying
-multiple approaches, repeating the process multiple times, and having
-some luck.
+In this example, we will show how the modular structure of an ecological network
+can be optimized. Finding the optimal modular structure can be a time-consuming
+process, as it relies on heuristic which are not guaranteed to converge to the
+global maximum. There is no elegant alternative to trying multiple approaches,
+repeating the process multiple times, and having some luck.
 
-We will use again the first network from the @HadfKras14 dataset in this
-example, which has a small number of species. For the first approach, we
-will generate random partitions of the species across 3 to 12 modules, and
-evaluate 20 replicate attempts for each of these combinations. The output
-we are interested in is the number of modules, and the overall modularity
-[@Barb07].
+```@example betadiv
+using EcologicalNetworks
+using Plots
+using EcologicalNetworksPlots
+```
 
-```julia
+For the first approach, we will generate random partitions of the species across
+3 to 12 modules, and evaluate 20 replicate attempts for each of these
+combinations. The output we are interested in is the number of modules, and the
+overall modularity.
+
+```@example modularity
 n = repeat(3:12, outer=20)
 m = Array{Dict}(undef, length(n))
 
@@ -25,63 +28,56 @@ for i in eachindex(n)
 end
 ```
 
+```@example modularity
+p1 = scatter(c, q, c=:grey, msw=0.0, leg=false, frame=:origin, grid=false)
+xaxis!(p1, "Number of modules")
+yaxis!(p1, "Modularity", (0, 0.5))
+```
+
 Now that we have the modular partition for every attempt, we can count the
 modules in it, and measure its modularity:
 
-```julia
+```@example modularity
 q = map(x -> Q(N,x), m);
 c = (m .|> values |> collect) .|> unique .|> length;
 ```
 
-The relationship between the two is represented in @fig:modularity. Out of
-the `j length(c)` attempts, we want to get the most modular one, *i.e.* the
-one with highest modularity. In some simple problems, there may be several
-partitions with the highest value, so we can either take the first, or one
-at random:
+Out of all attempts, we want to get the most modular one, *i.e.* the one with
+highest modularity. In some simple problems, there may be several partitions
+with the highest value, so we can either take the first, or one at random:
 
-```julia
+```@example modularity
 optimal = rand(findall(q.== maximum(q)));
 best_m = m[optimal];
 ```
 
-```julia; echo=false
-p1 = scatter(c, q, c=:grey, msw=0.0, leg=false, frame=:origin, grid=false)
-xaxis!(p1, "Number of modules")
-yaxis!(p1, "Modularity", (0, 0.5))
+This can be plotted using `EcologicalNetworksPlots`:
 
+```@example modularity
 I = initial(RandomInitialLayout, N)
 for step in 1:4000
   position!(ForceDirectedLayout(2.5), I, N)
 end
 p2 = plot(I, N, aspectratio=1)
 scatter!(p2, I, N, bipartite=true, nodefill=best_m, markercolor=:isolum)
-
-plot(p1, p2, size=(700,300))
-savefig("figures/modularity.pdf")
 ```
 
-![Left, relationship between the number of modules in the optimized partition and its modularity. Right, representation of the network where every node is colored according to the module it belongs to in the optimal partition.](figures/modularity.pdf){#fig:modularity}
[…]
