+++ docs/Project.toml
@@ -1,5 +1,6 @@
 [deps]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+EcologicalNetworksPlots = "9f7a259d-73a7-556d-a7a2-3eb122d3865b"
 Mangal = "b8b640a6-63d9-51e6-b784-5033db27bef2"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 
