+++ docs/src/examples/integration_mangal.md
@@ -11,7 +11,7 @@ using DataFrames
 using Plots
 ```
 
-We first retrieve relevant metadata for all networks archived on `mangal.io` using the `Mangal.jl` package. We count the number of species $S$ and the total number of interactions $L$ in each network, as well as their number of trophic interactions (predation and herbivory). We store these information in a data frame along with the networks' ID numbers.
+We first retrieve relevant metadata for all 1,386 networks archived on `mangal.io` using the `Mangal.jl` package. We count the number of species $S$ and the total number of interactions $L$ in each network, as well as their number of trophic interactions (predation and herbivory). We store these information in a data frame along with the networks' ID numbers, and print the first 5 elements. Due to the high number of networks we handle, note that this step might take some time to run.
 
 ```@example mangal
 number_of_networks = count(MangalNetwork)
@@ -23,10 +23,10 @@ mangal_networks = DataFrame(fill(Int64, 5),
                  number_of_networks)
 
 global cursor = 1
-@progress "Paging networks" for page in 1:number_of_pages
+for page in 1:number_of_pages
     global cursor
     networks_in_page = Mangal.networks("count" => count_per_page, "page" => page-1)
-    @progress "Counting items" for current_network in networks_in_page
+    for current_network in networks_in_page
         S = count(MangalNode, current_network)
         L = count(MangalInteraction, current_network)
         pred = count(MangalInteraction, current_network, "type" => "predation")
@@ -35,15 +35,19 @@ global cursor = 1
         cursor = cursor + 1
     end
 end
+
+first(mangal_networks, 5)
 ```
 
-We now have all the information we need to identify all food webs archived on `mangal.io`. Here we consider as food webs any ecological networks mainly composed of trophic interactions.
+We now have all the information we need to identify all food webs archived on `mangal.io`. Here we consider as food webs any ecological networks mainly composed of trophic interactions. We find that 259 networks meet this condition.  
 
 ```@example mangal
-mangal_foodwebs = mangal_networks[mangal_networks[!, :pred] .+ mangal_networks[!, :herb] ./ mangal_networks[!, :L] .> 0.5, :]
+foodwebs = mangal_networks[mangal_networks[!, :pred] .+ mangal_networks[!, :herb] ./ mangal_networks[!, :L] .> 0.5, :]
+
+first(foodwebs, 5)
 ```
 
-To analyse their properties, we first need to read all of these food webs using the `Mangal.jl` package, and then convert them to UnipartiteNetworks, when possible, using the `EcologicalNetworks.jl` package.
+To analyse their properties, we first need to read all of these food webs using the `Mangal.jl` package, and then convert them to UnipartiteNetworks, when possible, using the `EcologicalNetworks.jl` package. We end up with 241 unipartite food webs.
 
 ```@example mangal
 mangal_foodwebs = network.(foodwebs.id)
@@ -55,12 +59,13 @@ for i in eachindex(mangal_foodwebs)
         unipartite_foodweb = convert(UnipartiteNetwork, mangal_foodwebs[i])
         push!(unipartite_foodwebs, unipartite_foodweb)
     catch
-        println("Cannot convert mangal food web $(i) to a unipartite network")
     end
 end
+
+unipartite_foodwebs[1:5]
 ```
 
-We can then compute any measure supported by `EcologicalNetworks.jl` for UnipartiteNetworks. In this example, we compute species richness, connectance, nestedness, and modularity. To compute network modularity, we use 100 random species assignments in 3 to 15 groups as our starters, the BRIM algorithm to optimize the modularity for each of these random partitions, and retain the maximum value for each food web.
[…]
