### 2020-11-12 1496f6c
fix figure size
