### 2020-11-12 9f381f4
load the networks
