### 2020-11-12 21797b7
:bug: attempt to fix figure size
