### 2020-11-13 EcoJulia/EcologicalNetworks.jl#171 @FrancisBanville
Thank you for your suggestion @arbennett!  We added [an example](https://ecojulia.github.io/EcologicalNetworks.jl/dev/examples/integration_mangal/) inspired from the code example given in the paper, and will definitely add more examples in the upcoming months.
