### 2020-11-13 #2721 @FrancisBanville
Thanks for your review @arbennett!

We changed the defective link in the revised paper (branch `joss-article-review1`), and included a code example in the documentation of our package. We also recently changed the look of our conceptual figure to make it slightly more attractive.
