### 2020-11-13 EcoJulia/EcologicalNetworks.jl#171 @arbennett
Excellent! I just read through the example and think it's great! Closing.
