### 2020-11-18 #2721 @KristinaRiemer
Hi @arbennett. With regards to the authorship question, one of the authors had previously been the only contributor to the code base, while the submitting author and the third additional author had not. Based on the comments in [#166](https://github.com/EcoJulia/EcologicalNetworks.jl/issues/166) from the authors, all of their contributions seem more than sufficient for authorship as we require at JOSS. Does that make sense?

@kellykochanski it seems that the authors have addressed your feedback in [#167](https://github.com/EcoJulia/EcologicalNetworks.jl/issues/167) and [#168](https://github.com/EcoJulia/EcologicalNetworks.jl/issues/168), if you are able to follow up with those and continue your review?

Please let me know if anyone has additional questions!
