### 2020-11-27 f6c8b35
📖 example on extinctions
