### 2020-11-27 779f6f8
📖 extinctions example
