### 2020-11-27 4bf2ddf
⚡ improve the interactions function
