### 2020-11-27 927e2da
⚡ speedup interactions by pre-allocating the array
