### 2020-11-27 6790987
📖 use a modular network for the modularity page
