### 2020-11-27 7336853
📊 fix the margins in the example plots
