+++ .github/workflows/Documentation.yml
@@ -4,7 +4,7 @@ on:
   release:
   push:
     branches:
-      - master
+      - main
     tags: '*'
   pull_request:
 
