+++ src/types/constructors.jl
@@ -1,16 +1,3 @@
-function BipartiteQuantitativeNetwork(A::Matrix{IT}) where {IT <: Number}
-  check_bipartiteness(A)
-  T = "t".*string.(1:size(A,1))
-  B = "b".*string.(1:size(A,2))
-  BipartiteQuantitativeNetwork{IT,eltype(T)}(A, T, B)
-end
-
-function BipartiteQuantitativeNetwork(A::Matrix{IT}, T::Vector{NT}, B::Vector{NT}) where {IT<:Number,NT}
-  check_species_validity(NT)
-  check_bipartiteness(A, T, B)
-  BipartiteQuantitativeNetwork{IT,NT}(A, T, B)
-end
-
 function UnipartiteQuantitativeNetwork(A::Matrix{IT}) where {IT<:Number}
   check_unipartiteness(A)
   S = "s".*string.(1:size(A,1))
@@ -22,8 +9,6 @@ function UnipartiteQuantitativeNetwork(A::Matrix{IT}, S::Vector{NT}) where {IT<:
   check_unipartiteness(A, S)
   UnipartiteQuantitativeNetwork{IT,NT}(A, S)
 end
-
-
 
 function UnipartiteProbabilisticNetwork(A::Matrix{IT}) where {IT<:AbstractFloat}
   check_unipartiteness(A)
