+++ src/types/utilities.jl
@@ -14,9 +14,9 @@ check_species_validity(::Type{T}) where {T <: String} = nothing
 function show(io::IO, N::AbstractEcologicalNetwork)
   p = typeof(N) <: AbstractBipartiteNetwork ? "bipartite" : "unipartite"
   t = ""
-  t = typeof(N) <: ProbabilisticNetwork ? "probabilistic" : t
-  t = typeof(N) <: QuantitativeNetwork ? "quantitative" : t
-  print(io, "$(richness(N; dims=1))×$(richness(N; dims=2)) $(p) $(t) ecological network $(eltype(N)) (L: $(links(N)))")
+  t = typeof(N) <: ProbabilisticNetwork ? " probabilistic" : t
+  t = typeof(N) <: QuantitativeNetwork ? " quantitative" : t
+  print(io, "$(richness(N; dims=1))×$(richness(N; dims=2)) $(p)$(t) ecological network $(eltype(N)) (L: $(links(N)))")
 end
 
 """
@@ -93,8 +93,8 @@ but is exported because it may be of general use.
 """
 function has_interaction(N::AbstractEcologicalNetwork, i::Int64, j::Int64)
   # We need to make sure that the interaction is somewhere within the network
-  @assert i <= size(N.A, 1)
-  @assert j <= size(N.A, 2)
+  @assert i <= size(N.edges, 1)
+  @assert j <= size(N.edges, 2)
   # This should be reasonably general...
   return !iszero(N[i,j])
 end
@@ -105,7 +105,7 @@ end
 Modifies the network so that its diagonal is set to the appropriate zero.
 """
 function nodiagonal!(N::AbstractUnipartiteNetwork)
-  N.A[diagind(N.A)] .= zero(eltype(N.A))
+  N.edges[diagind(N.edges)] .= zero(eltype(N.edges))
 end
 
 """
@@ -147,7 +147,7 @@ end
 Return the size of the adjacency matrix of an `AbstractEcologicalNetwork` object.
 """
 function size(N::AbstractEcologicalNetwork)
-  size(N.A)
+  size(N.edges)
 end
 
 
@@ -156,9 +156,9 @@ Creates a copy of a network -- this returns an object with the same type, and
 the same content.
 """
 function copy(N::AbstractEcologicalNetwork)
-  a = copy(N.A)
+  edges = copy(N.edges)
   sp = copy.(species_objects(N))
-  return typeof(N)(a, sp...)
+  return typeof(N)(edges, sp...)
 end
 
 
@@ -168,7 +168,7 @@ end
 Get the value of an interaction based on the position of the species.
 """
 function getindex(N::AbstractEcologicalNetwork, i::T, j::T) where {T <: Int}
-  return N.A[i,j]
+  return N.edges[i,j]
 end
 
 """
@@ -177,7 +177,7 @@ end
 Get the value of an interaction based on the position of the species.
 """
 function getindex(N::AbstractEcologicalNetwork, i::T) where {T <: Int}
-  return N.A[i]
+  return N.edges[i]
 end
 
 
@@ -187,7 +187,7 @@ end
 Get the value of an interaction based on the position of the species.
 """
 function getindex(N::AbstractEcologicalNetwork, ::Colon, j::T) where {T <: Int}
-  return N.A[:,j]
+  return N.edges[:,j]
 end
 
 """
@@ -196,7 +196,7 @@ end
 Get the value of an interaction based on the position of the species.
 """
 function getindex(N::AbstractEcologicalNetwork, i::T, ::Colon) where {T <: Int}
-  return N.A[i,:]
[…]
