+++ src/types/declarations.jl
@@ -101,30 +101,28 @@ A bipartite probabilistic network is a matrix of floating point numbers, all of
 which must be between 0 and 1.
 """
 mutable struct BipartiteProbabilisticNetwork{IT <: AbstractFloat, ST} <: AbstractBipartiteNetwork
-    edges::SparseMatrixCSC{IT,Int64}
+    edges::SparseMatrixCSC{IT}
     T::Vector{ST}
     B::Vector{ST}
-    function BipartiteProbabilisticNetwork{IT,ST}(edges::SparseMatrixCSC{IT,Int64}, T::Vector{ST}, B::Vector{ST}) where {IT <: AbstractFloat, ST}
+    function BipartiteProbabilisticNetwork(edges, T, B)
         dropzeros!(edges)
+        allunique(T) || throw(ArgumentError("All top-level species must be unique"))
+        allunique(B) || throw(ArgumentError("All bottom-level species must be unique"))
+        allunique(vcat(B,T)) || throw(ArgumentError("Bipartite networks cannot share species across levels"))
         check_probability_values(edges)
-        new{IT,ST}(edges, T, B)
+        check_species_validity(eltype(T))
+        check_species_validity(eltype(B))
+        check_bipartiteness(edges, T, B)
+        new{eltype(edges),eltype(T)}(edges, T, B)
+    end
+    function BipartiteProbabilisticNetwork{IT,ST}(edges::SparseMatrixCSC{IT}, T::Vector{ST}, B::Vector{ST}) where {IT <: Real, ST}
+        return BipartiteProbabilisticNetwork(edges, T, B)
     end
 end
 
 function BipartiteProbabilisticNetwork(A::Array{IT,2}, T::Union{Vector{TT},Nothing}=nothing, B::Union{Vector{TT},Nothing}=nothing) where {IT <: AbstractFloat, TT}
-    if isnothing(B)
-        B = "b".*string.(1:size(A, 2))
-    else
-        check_species_validity(TT)
-    end
-    if isnothing(T)
-        T = "t".*string.(1:size(A, 1))
-    else
-        check_species_validity(TT)
-    end
-    allunique(T) || throw(ArgumentError("All top-level species must be unique"))
-    allunique(B) || throw(ArgumentError("All bottom-level species must be unique"))
-    allunique(vcat(B,T)) || throw(ArgumentError("Bipartite networks cannot share species across levels"))
+    isnothing(B) ? (B = "b".*string.(1:size(A, 2))) : check_species_validity(TT)
+    isnothing(T) ? (T = "t".*string.(1:size(A, 1))) : check_species_validity(TT)
     isequal(length(T))(size(A,1)) || throw(ArgumentError("The matrix has the wrong number of top-level species"))
     isequal(length(B))(size(A,2)) || throw(ArgumentError("The matrix has the wrong number of bottom-level species"))
     return BipartiteProbabilisticNetwork(sparse(A), T, B)
@@ -135,32 +133,30 @@ A bipartite quantitative network is matrix of numbers. It is assumed that the
 interaction strength are *positive*.
 """
 mutable struct BipartiteQuantitativeNetwork{IT <: Number, ST} <: AbstractBipartiteNetwork
-    edges::SparseMatrixCSC{IT,Int64}
+    edges::SparseMatrixCSC{IT}
     T::Vector{ST}
     B::Vector{ST}
-    function BipartiteQuantitativeNetwork{IT, NT}(edges::SparseMatrixCSC{IT,Int64}, T::Vector{NT}, B::Vector{NT}) where {IT <: Number, NT}
+    function BipartiteQuantitativeNetwork(edges, T, B)
         dropzeros!(edges)
[…]
