+++ src/links/degree.jl
@@ -64,7 +64,7 @@ end
     degree_out_var(N::ProbabilisticNetwork)
 """
 function degree_out_var(N::ProbabilisticNetwork)
-  d_o_v = mapslices(a_var, N.A; dims=2)
+  d_o_v = vec(mapslices(a_var, convert(Matrix, N.edges); dims=2))
   return Dict(zip(species(N; dims=1), d_o_v))
 end
 
@@ -74,7 +74,7 @@ end
     degree_in_var(N::ProbabilisticNetwork)
 """
 function degree_in_var(N::ProbabilisticNetwork)
-  d_i_v = mapslices(a_var, N.A; dims=1)'
+  d_i_v = vec(mapslices(a_var, convert(Matrix, N.edges); dims=1))
   return Dict(zip(species(N; dims=2), d_i_v))
 end
 
