### 2020-11-28 627fdb7
🎉 clean definition of interfaces
