+++ src/types/utilities.jl
@@ -1,4 +1,4 @@
-import Base: getindex, setindex!, permutedims, permutedims!, size, copy, !, show, +, inv, similar
+import Base: getindex, setindex!, permutedims, permutedims!, copy, !, show, +, inv
 
 function check_species_validity(::Type{T}) where {T <: Any}
   throw(ArgumentError("The type $(T) is not an allowed species type"))
@@ -140,17 +140,6 @@ Does nothing.
 function nodiagonal!(N::AbstractBipartiteNetwork)
 end
 
-
-"""
-    size(N::AbstractEcologicalNetwork)
-
-Return the size of the adjacency matrix of an `AbstractEcologicalNetwork` object.
-"""
-function size(N::AbstractEcologicalNetwork)
-  size(N.edges)
-end
-
-
 """
 Creates a copy of a network -- this returns an object with the same type, and
 the same content.
@@ -159,195 +148,6 @@ function copy(N::AbstractEcologicalNetwork)
   edges = copy(N.edges)
   sp = copy.(species_objects(N))
   return typeof(N)(edges, sp...)
-end
-
-
-"""
-    getindex(N::AbstractEcologicalNetwork, i::T, j::T)
-
-Get the value of an interaction based on the position of the species.
-"""
-function getindex(N::AbstractEcologicalNetwork, i::T, j::T) where {T <: Int}
-  return N.edges[i,j]
-end
-
-"""
-    getindex(N::AbstractEcologicalNetwork, i::T)
-
-Get the value of an interaction based on the position of the species.
-"""
-function getindex(N::AbstractEcologicalNetwork, i::T) where {T <: Int}
-  return N.edges[i]
-end
-
-
-"""
-    getindex(N::AbstractEcologicalNetwork, ::Colon, j::T)
-
-Get the value of an interaction based on the position of the species.
-"""
-function getindex(N::AbstractEcologicalNetwork, ::Colon, j::T) where {T <: Int}
-  return N.edges[:,j]
-end
-
-"""
-    getindex(N::AbstractEcologicalNetwork, i::T, ::Colon)
-
-Get the value of an interaction based on the position of the species.
-"""
-function getindex(N::AbstractEcologicalNetwork, i::T, ::Colon) where {T <: Int}
-  return N.edges[i,:]
-end
-
-"""
-    getindex{T}(N::AbstractEcologicalNetwork, s1::T, s2::T)
-
-Get the value of an interaction based on the *name* of the species. This is the
-recommended way to look for things in a network.
-"""
-function getindex(N::AbstractEcologicalNetwork, s1::T, s2::T) where {T}
-  @assert T == last(eltype(N))
-  @assert s1 ∈ species(N; dims=1)
-  @assert s2 ∈ species(N; dims=2)
-  s1_pos = something(findfirst(isequal(s1), species(N; dims=1)), 0)
-  s2_pos = something(findfirst(isequal(s2), species(N; dims=2)), 0)
-  return N[s1_pos, s2_pos]
-end
-
-
-"""
-    getindex{T}(N::AbstractBipartiteNetwork, ::Colon, sp::T)
-
-Gets the predecessors (*i.e.* species that interacts with / consume) of a focal
-species. This returns the list of species as a `Set` object, in which ordering
-is unimportant.
-"""
-function getindex(N::AbstractEcologicalNetwork, ::Colon, sp::T) where {T}
-  @assert T == last(eltype(N))
-  @assert sp ∈ species(N; dims=2)
[…]
