+++ src/interfaces/abstractarray.jl
@@ -1,0 +1,221 @@
+function Base.length(N::T) where {T <: AbstractEcologicalNetwork}
+    return count(!iszero, N.edges)
+end
+
+"""
+    similar(N::T) where {T <: AbstractEcologicalNetwork}
+
+Returns a network with the same species, but an empty interaction matrix. This
+is useful if you want to generate a "blank slate" for some analyses.
+"""
+function Base.similar(N::T) where {T <: AbstractEcologicalNetwork}
+   Y = sparse(zeros(first(eltype(N)), size(N)))
+   return T(Y, species_objects(N)...)
+end
+
+"""
+    size(N::AbstractEcologicalNetwork)
+
+Return the size of the adjacency matrix of an `AbstractEcologicalNetwork` object.
+"""
+function Base.size(N::AbstractEcologicalNetwork)
+  size(N.edges)
+end
+
+"""
+    size(N::AbstractEcologicalNetwork, i::Int64)
+
+Return the size of the adjacency matrix of an `AbstractEcologicalNetwork` object.
+"""
+function Base.size(N::AbstractEcologicalNetwork, i::Int64)
+  size(N.edges, i)
+end
+
+
+"""
+    getindex(N::AbstractEcologicalNetwork, i::T, j::T)
+
+Get the value of an interaction based on the position of the species.
+"""
+function Base.getindex(N::AbstractEcologicalNetwork, i::T, j::T) where {T <: Int}
+  return N.edges[i,j]
+end
+
+"""
+    getindex(N::AbstractEcologicalNetwork, i::T)
+
+Get the value of an interaction based on the position of the species.
+"""
+function Base.getindex(N::AbstractEcologicalNetwork, i::T) where {T <: Int}
+  return N.edges[i]
+end
+
+
+"""
+    getindex(N::AbstractEcologicalNetwork, ::Colon, j::T)
+
+Get the value of an interaction based on the position of the species.
+"""
+function Base.getindex(N::AbstractEcologicalNetwork, ::Colon, j::T) where {T <: Int}
+  return N.edges[:,j]
+end
+
+"""
+    getindex(N::AbstractEcologicalNetwork, i::T, ::Colon)
+
+Get the value of an interaction based on the position of the species.
+"""
+function Base.getindex(N::AbstractEcologicalNetwork, i::T, ::Colon) where {T <: Int}
+  return N.edges[i,:]
+end
+
+"""
+    getindex{T}(N::AbstractEcologicalNetwork, s1::T, s2::T)
+
+Get the value of an interaction based on the *name* of the species. This is the
+recommended way to look for things in a network.
+"""
+function Base.getindex(N::AbstractEcologicalNetwork, s1::T, s2::T) where {T}
+  @assert T == last(eltype(N))
+  @assert s1 ∈ species(N; dims=1)
+  @assert s2 ∈ species(N; dims=2)
+  s1_pos = something(findfirst(isequal(s1), species(N; dims=1)), 0)
+  s2_pos = something(findfirst(isequal(s2), species(N; dims=2)), 0)
+  return N[s1_pos, s2_pos]
+end
+
+
+"""
+    getindex{T}(N::AbstractBipartiteNetwork, ::Colon, sp::T)
+
+Gets the predecessors (*i.e.* species that interacts with / consume) of a focal
+species. This returns the list of species as a `Set` object, in which ordering
+is unimportant.
+"""
+function Base.getindex(N::AbstractEcologicalNetwork, ::Colon, sp::T) where {T}
[…]
