+++ src/EcologicalNetworks.jl
@@ -50,15 +50,14 @@ function __init__()
    end
 end
 
+# Interfaces for networks
+include(joinpath(".", "interfaces/abstractarray.jl"))
+include(joinpath(".", "interfaces/iteration.jl"))
+
 # General useful manipulations
 include(joinpath(".", "types/utilities.jl"))
 include(joinpath(".", "types/comparisons.jl"))
 export species, interactions, has_interaction, richness, nodiagonal, nodiagonal!
-
-# Interfaces for networks
-include(joinpath(".", "interfaces/iteration.jl"))
-include(joinpath(".", "interfaces/abstractarray.jl"))
-include(joinpath(".", "interfaces/indexing.jl"))
 
 # Degree
 include(joinpath(".", "links/degree.jl"))
