### 2020-11-28 9d6dd03
🧹 separation of network interface from the type business
