### 2020-11-28 64c3b13
🧹 clarify name of probability functions
