### 2020-11-28 dc36849
⚡ cleanup the multiplicative variance function
