### 2020-11-28 c7a6271
🔧 fix the simplify! code for unipartite networks
