+++ src/links/degenerate.jl
@@ -10,11 +10,11 @@ function isdegenerate(N::AbstractEcologicalNetwork)
 end
 
 """
-    simplify{T<:AbstractBipartiteNetwork}(N::T)
+    simplify(N::T) where {T<:AbstractEcoloigcalNetwork}
 
 Returns a new network in which species with no interactions have been removed.
 """
-function simplify(N::T) where {T<:AbstractBipartiteNetwork}
+function simplify(N::T) where {T<:AbstractEcologicalNetwork}
     Y = copy(N)
     simplify!(Y)
     return Y
@@ -38,25 +38,58 @@ function simplify!(N::T) where {T<:AbstractBipartiteNetwork}
 end
 
 """
-    simplify(N::AbstractUnipartiteNetwork)
-
-Returns a new network in which species with no interactions have been removed.
-"""
-function simplify(N::T) where {T <: AbstractUnipartiteNetwork}
-    Y = copy(N)
-    simplify!(Y)
-    return Y
-end
-
-"""
-    simplify!(N::AbstractUnipartiteNetwork)
+    simplify!(N::UnipartiteNetwork)
 
 Modifies the network to drop all species without an interaction.
 """
-function simplify!(N::T) where {T <: AbstractUnipartiteNetwork}
-    d = degree(N)
-    s = species(N)
-    positions = filter(i -> d[s[i]] > zero(eltype(N)[1]), 1:richness(N))
-    N.S = N.S[positions]
-    N.A = N.A[positions, positions]
+function simplify!(N::T) where {T <: UnipartiteNetwork}
+    isdegenerate(N) || return nothing
+    int = interactions(N)
+    from = [i.from for i in int]
+    to = [i.to for i in int]
+    sp = unique(vcat(from, to))
+    I = indexin(from, sp)
+    J = indexin(to, sp)
+    V = true
+    N.S = sp
+    N.edges = sparse(I, J, V, length(sp), length(sp))
+    return nothing
+end
+
+"""
+    simplify!(N::UnipartiteNetwork)
+
+Modifies the network to drop all species without an interaction.
+"""
+function simplify!(N::T) where {T <: UnipartiteProbabilisticNetwork}
+    isdegenerate(N) || return nothing
+    int = interactions(N)
+    from = [i.from for i in int]
+    to = [i.to for i in int]
+    sp = unique(vcat(from, to))
+    I = indexin(from, sp)
+    J = indexin(to, sp)
+    V = [i.probability for i in int]
+    N.S = sp
+    N.edges = sparse(I, J, V, length(sp), length(sp))
+    return nothing
+end
+
+"""
+    simplify!(N::UnipartiteNetwork)
+
+Modifies the network to drop all species without an interaction.
+"""
+function simplify!(N::T) where {T <: UnipartiteQuantitativeNetwork}
+    isdegenerate(N) || return nothing
+    int = interactions(N)
+    from = [i.from for i in int]
+    to = [i.to for i in int]
+    sp = unique(vcat(from, to))
+    I = indexin(from, sp)
+    J = indexin(to, sp)
+    V = [i.strength for i in int]
+    N.S = sp
+    N.edges = sparse(I, J, V, length(sp), length(sp))
+    return nothing
 end
