### 2020-11-28 bf839ea
🎉 update simplify! code
