+++ src/links/degenerate.jl
@@ -21,20 +21,57 @@ function simplify(N::T) where {T<:AbstractEcologicalNetwork}
 end
 
 """
-    simplify!{T<:AbstractBipartiteNetwork}(N::T)
+    simplify!{T<:BipartiteNetwork}(N::T)
 
 Returns a new network in which species with no interactions have been removed.
 """
-function simplify!(N::T) where {T<:AbstractBipartiteNetwork}
-    d1 = degree(N; dims=1)
-    d2 = degree(N; dims=2)
-    s1 = species(N; dims=1)
-    s2 = species(N; dims=2)
-    p1 = filter(i -> d1[s1[i]] > zero(eltype(N)[1]), 1:richness(N; dims=1))
-    p2 = filter(i -> d2[s2[i]] > zero(eltype(N)[1]), 1:richness(N; dims=2))
-    N.T = N.T[p1]
-    N.B = N.B[p2]
-    N.A = N.A[p1,p2]
+function simplify!(N::T) where {T<:BipartiteNetwork}
+    isdegenerate(N) || return nothing 
+    int = interactions(N)
+    from = [i.from for i in int]
+    to = [i.to for i in int]
+    N.T = unique(from)
+    N.B = unique(to)
+    I = indexin(from, N.T)
+    J = indexin(to, N.B)
+    N.edges = sparse(I, J, true, length(N.T), length(N.B))
+    return nothing
+end
+
+"""
+    simplify!{T<:BipartiteNetwork}(N::T)
+
+Returns a new network in which species with no interactions have been removed.
+"""
+function simplify!(N::T) where {T<:BipartiteProbabilisticNetwork}
+    isdegenerate(N) || return nothing 
+    int = interactions(N)
+    from = [i.from for i in int]
+    to = [i.to for i in int]
+    N.T = unique(from)
+    N.B = unique(to)
+    I = indexin(from, N.T)
+    J = indexin(to, N.B)
+    N.edges = sparse(I, J, [i.probability for i in int], length(N.T), length(N.B))
+    return nothing
+end
+
+"""
+    simplify!{T<:BipartiteNetwork}(N::T)
+
+Returns a new network in which species with no interactions have been removed.
+"""
+function simplify!(N::T) where {T<:BipartiteQuantitativeNetwork}
+    isdegenerate(N) || return nothing 
+    int = interactions(N)
+    from = [i.from for i in int]
+    to = [i.to for i in int]
+    N.T = unique(from)
+    N.B = unique(to)
+    I = indexin(from, N.T)
+    J = indexin(to, N.B)
+    N.edges = sparse(I, J, [i.strength for i in int], length(N.T), length(N.B))
+    return nothing
 end
 
 """
