### 2020-11-28 9ec14ba
🎉 fix rand (in a not quite optimal way)
