+++ src/rand/draws.jl
@@ -1,5 +1,3 @@
-import Base.rand
-
 """
     rand(N::ProbabilisticNetwork)
 
@@ -10,12 +8,21 @@ interactions.
 
 #### References
 
-Poisot, T., Cirtwill, A.R., Cazelles, K., Gravel, D., Fortin, M.-J., Stouffer, D.B., 2016. The structure of probabilistic networks. Methods in Ecology and Evolution 7, 303–312. https://doi.org/10.1111/2041-210X.12468
+- Poisot, T., Cirtwill, A.R., Cazelles, K., Gravel, D., Fortin, M.-J., Stouffer,
+  D.B., 2016. The structure of probabilistic networks. Methods in Ecology and
+  Evolution 7, 303–312. https://doi.org/10.1111/2041-210X.12468
 """
-function rand(N::ProbabilisticNetwork)
+function Base.rand(N::ProbabilisticNetwork)
     # Get the correct network type
     newtype = typeof(N) <: AbstractUnipartiteNetwork ? UnipartiteNetwork : BipartiteNetwork
-    return newtype(rand(Float64, size(N)).<=N.A, species_objects(N)...)
+    ed = spzeros(Bool, size(N.edges)...)
+    K = newtype(ed, EcologicalNetworks.species_objects(N)...)
+    r = rand(length(N))
+    for (i,int) in enumerate(interactions(N))
+        (r[i] ≤ int.probability) && (K[int.from, int.to] = true )
+    end
+    dropzeros!(K.edges)
+    return K
 end
 
 """
@@ -26,9 +33,11 @@ network.
 
 #### References
 
-Poisot, T., Cirtwill, A.R., Cazelles, K., Gravel, D., Fortin, M.-J., Stouffer, D.B., 2016. The structure of probabilistic networks. Methods in Ecology and Evolution 7, 303–312. https://doi.org/10.1111/2041-210X.12468
+- Poisot, T., Cirtwill, A.R., Cazelles, K., Gravel, D., Fortin, M.-J., Stouffer,
+  D.B., 2016. The structure of probabilistic networks. Methods in Ecology and
+  Evolution 7, 303–312. https://doi.org/10.1111/2041-210X.12468
 """
-function rand(N::ProbabilisticNetwork, n::T) where {T<:Integer}
+function Base.rand(N::ProbabilisticNetwork, n::T) where {T<:Integer}
     @assert n > 0
     return map(x -> rand(N), 1:n)
 end
@@ -41,9 +50,11 @@ network, and returns them as a matrix.
 
 #### References
 
-Poisot, T., Cirtwill, A.R., Cazelles, K., Gravel, D., Fortin, M.-J., Stouffer, D.B., 2016. The structure of probabilistic networks. Methods in Ecology and Evolution 7, 303–312. https://doi.org/10.1111/2041-210X.12468
+- Poisot, T., Cirtwill, A.R., Cazelles, K., Gravel, D., Fortin, M.-J., Stouffer,
+  D.B., 2016. The structure of probabilistic networks. Methods in Ecology and
+  Evolution 7, 303–312. https://doi.org/10.1111/2041-210X.12468
 """
-function rand(N::ProbabilisticNetwork, S::Tuple{Int64,Int64})
+function Base.rand(N::ProbabilisticNetwork, S::Tuple{Int64,Int64})
     @assert minimum(S) > 0
-    return reshape(map(x -> rand(N), 1:prod(S)), S)
+    return reshape(rand(N, prod(S)), S)
 end
