### 2020-11-28 5002164
⚡ faster rand
