+++ docs/src/interface/types.md
@@ -70,7 +70,7 @@ AbstractEcologicalNetwork
 ```
 
 The type of nodes that are allowed is determined by the *non-exported*
-`EcologicalNetworks.check_species_validity` function. To allow an additional type of
+`EcologicalNetworks._check_species_validity` function. To allow an additional type of
 node, you can write the following:
 
 ~~~ julia
@@ -81,12 +81,12 @@ struct Foo
 end
 
 import EcologicalNetworks
-function EcologicalNetworks.check_species_validity(::Type{Foo})
+function EcologicalNetworks._check_species_validity(::Type{Foo})
 end
 ~~~
 
 Note that **integers are never valid species identifiers**. By default, `String`
-and `Symbol` are used. The function `check_species_validity` should do *nothing*
+and `Symbol` are used. The function `_check_species_validity` should do *nothing*
 for an accepted type (and it will throw an error for any other type).
 
 ### By partiteness
