+++ src/community/overlap.jl
@@ -136,7 +136,7 @@ https://doi.org/10.1016/j.ecoinf.2015.09.013
 function EAJS(N::T; dims::Union{Nothing,Integer}=nothing) where {T <: UnipartiteNetwork}
 
     sp = shortest_path(N)
-    Y = UnipartiteNetwork(sp.>0, species_objects(N)...)
+    Y = UnipartiteNetwork(sp.>0, _species_objects(N)...)
 
     if dims !== nothing
         dims == 1 || dims == 2 || throw(ArgumentError("dims can only be nothing, 1, or 2 -- you used $(dims)"))
