+++ docs/make.jl
@@ -13,9 +13,9 @@ makedocs(
         "Interface" => [
             "Types" => "interface/types.md",
             "Conversions" => "interface/conversions.md",
-            "Core functions" => "interface/highlevel.md"
+            "Core functions" => "interface/highlevel.md",
             "AbstractArray" => "interface/abstractarray.md",
-            "Iterate" => "interface/iterate.md",
+            "Iterate" => "interface/iterate.md"
         ],
         "Examples" => [
             "Modularity" => "examples/modularity.md",
