### 2020-11-29 3a4be41
🎨 update the show method
