### 2020-11-30 b0b9cf3
sparse motif permutations
