+++ src/community/motifs.jl
@@ -52,25 +52,27 @@ Internal function
 Returns all permutations of the adjacency matrix of a motif.
 """
 function _permute_motif(m::T) where {T<:BipartiteNetwork}
-    perm = Array{Bool,2}[]
-    for ts in permutations(1:richness(m; dims=1)), bs in permutations(1:richness(m; dims=2))
-        push!(perm, m.A[ts,bs])
+    sp_permutations_T = permutations(1:richness(m; dims=1))
+    sp_permutations_B = permutations(1:richness(m; dims=2))
+    perm = fill(copy(m.edges), length(sp_permutations_T)*length(sp_permutations_B))
+    for (i, t_order) in enumerate(sp_permutations_T), (j, b_order) in enumerate(sp_permutations_B)
+        perm[(i-1)*j+j] = perm[(i-1)*j+j][t_order, b_order]
     end
     return unique(perm)
 end
 
-function inner_find_motif(N::T1, m::T2) where {T1<:UnipartiteNetwork,T2<:UnipartiteNetwork}
+function _inner_find_motif(N::T1, m::T2) where {T1<:UnipartiteNetwork,T2<:UnipartiteNetwork}
     motif_permutations = _permute_motif(m)
     matching_species = []
     for species_combination in combinations(species(N), richness(m))
-        if N[species_combination].A ∈ motif_permutations
+        if N.edges[species_combination] ∈ motif_permutations
             push!(matching_species, (species_combination,))
         end
     end
     return matching_species
 end
 
-function inner_find_motif(N::T1, m::T2) where {T1<:BipartiteNetwork,T2<:BipartiteNetwork}
+function _inner_find_motif(N::T1, m::T2) where {T1<:BipartiteNetwork,T2<:BipartiteNetwork}
     motif_permutations = _permute_motif(m)
     matching_species = []
     top_combinations = combinations(species(N; dims=1), richness(m; dims=1))
@@ -89,7 +91,7 @@ function inner_find_motif(N::T1, m::T2) where {T1<:BipartiteNetwork,T2<:Bipartit
             for i in 1:length(bottom_species)
                 bot_sp_pos[i] = p2[bottom_species[i]]
             end
-            if N.A[top_sp_pos, bot_sp_pos] ∈ motif_permutations
+            if N.edges[top_sp_pos, bot_sp_pos] ∈ motif_permutations
                 push!(matching_species, (top_species, bottom_species))
             end
         end
@@ -97,14 +99,14 @@ function inner_find_motif(N::T1, m::T2) where {T1<:BipartiteNetwork,T2<:Bipartit
     return matching_species
 end
 
-function inner_find_motif(N::T1, m::T2) where {T1<:UnipartiteProbabilisticNetwork, T2<:UnipartiteNetwork}
+function _inner_find_motif(N::T1, m::T2) where {T1<:UnipartiteProbabilisticNetwork, T2<:UnipartiteNetwork}
     motif_permutations = _permute_motif(m)
     all_combinations = []
     for species_combination in combinations(species(N), richness(m))
         isg = N[species_combination]
         motif_mean, motif_var = 0.0, 0.0
         for perm in motif_permutations
-            imat = zeros(eltype(N.A), size(m))
+            imat = zeros(eltype(N.edges), size(m))
             for i in eachindex(imat)
                 imat[i] = (perm[i] ? 2.0*isg[i] : 1.0)-isg[i]
             end
[…]
