### 2020-11-30 8db4775
🎉 spectral radius
