+++ src/betadiversity/operations.jl
@@ -1,14 +1,10 @@
-import Base.union
-import Base.intersect
-import Base.setdiff
-
 """
     union(X::T, Y::T) where {T<:BipartiteNetwork}
 
 Union of two bipartite networks -- interactions *and* species which are present
 in either networks are also present in the final network.
 """
-function union(X::T, Y::T) where {T<:BipartiteNetwork}
+function Base.union(X::T, Y::T) where {T<:BipartiteNetwork}
     new_t = union(species(X; dims=1), species(Y; dims=1))
     new_b = union(species(X; dims=2), species(Y; dims=2))
     new_a = spzeros(_interaction_type(X), length(new_t), length(new_b))
@@ -26,7 +22,7 @@ and interactions* common to both networks. This can result in species being
 disconnected, if they are found in both networks but with no operations in
 common.
 """
-function intersect(X::T, Y::T) where {T<:BipartiteNetwork}
+function Base.intersect(X::T, Y::T) where {T<:BipartiteNetwork}
     new_t = intersect(species(X; dims=1), species(Y; dims=1))
     new_b = intersect(species(X; dims=2), species(Y; dims=2))
     new_a = spzeros(_interaction_type(X), length(new_t), length(new_b))
@@ -42,23 +38,11 @@ end
 Union of two unipartite networks -- interactions *and* species which are present
 in either networks are also present in the final network.
 """
-function union(X::T, Y::T) where {T<:UnipartiteNetwork}
+function Base.union(X::T, Y::T) where {T<:UnipartiteNetwork}
     new_s = union(species(X), species(Y))
-    new_a = zeros(eltype(X.A), (length(new_s), length(new_s)))
-    for si in eachindex(new_s), sj in eachindex(new_s)
-        s1, s2 = new_s[si], new_s[sj]
-        in_x, in_y = false, false
-        if s1 ∈ species(X)
-            if s2 ∈ species(X)
-                in_x = has_interaction(X, s1, s2)
-            end
-        end
-        if s1 ∈ species(Y)
-            if s2 ∈ species(Y)
-                in_y = has_interaction(Y, s1, s2)
-            end
-        end
-        new_a[si,sj] = (in_x | in_y)
+    new_a = spzeros(_interaction_type(X), length(new_s), length(new_s))
+    for common in union(interactions(X), interactions(Y))
+        new_a[common.from,common.to] = true
     end
     return UnipartiteNetwork(new_a, new_s)
 end
@@ -71,15 +55,11 @@ Intersect between two unipartite networks. The resulting network has the
 being disconnected, if they are found in both networks but with no operations in
 common.
 """
-function intersect(X::T, Y::T) where {T<:UnipartiteNetwork}
+function Base.intersect(X::T, Y::T) where {T<:UnipartiteNetwork}
     new_s = intersect(species(X), species(Y))
-    new_a = zeros(eltype(X.A), (length(new_s), length(new_s)))
-    for si in eachindex(new_s), sj in eachindex(new_s)
-        s1, s2 = new_s[si], new_s[sj]
-        in_x, in_y = false, false
-        in_x = has_interaction(X, s1, s2)
-        in_y = has_interaction(Y, s1, s2)
-        new_a[si,sj] = (in_x & in_y)
+    new_a = spzeros(_interaction_type(X), length(new_s), length(new_s))
+    for common in intersect(interactions(X), interactions(Y))
+        new_a[common.from,common.to] = true
     end
     return UnipartiteNetwork(new_a, new_s)
 end
@@ -91,7 +71,7 @@ Difference between two bipartite networks. This operation is *sensitive* to the
 order of arguments, as the resulting network will have the species present in
[…]
