+++ src/betadiversity/operations.jl
@@ -8,10 +8,11 @@ function Base.union(X::T, Y::T) where {T<:BipartiteNetwork}
     new_t = union(species(X; dims=1), species(Y; dims=1))
     new_b = union(species(X; dims=2), species(Y; dims=2))
     new_a = spzeros(_interaction_type(X), length(new_t), length(new_b))
+    Z = BipartiteNetwork(new_a, new_t, new_b)
     for common in union(interactions(X), interactions(Y))
-        new_a[common.from,common.to] = true
+        Z[common.from,common.to] = true
     end
-    return BipartiteNetwork(new_a, new_t, new_b)
+    return Z
 end
 
 """
@@ -26,10 +27,11 @@ function Base.intersect(X::T, Y::T) where {T<:BipartiteNetwork}
     new_t = intersect(species(X; dims=1), species(Y; dims=1))
     new_b = intersect(species(X; dims=2), species(Y; dims=2))
     new_a = spzeros(_interaction_type(X), length(new_t), length(new_b))
+    Z = BipartiteNetwork(new_a, new_t, new_b)
     for common in intersect(interactions(X), interactions(Y))
-        new_a[common.from,common.to] = true
+        Z[common.from,common.to] = true
     end
-    return BipartiteNetwork(new_a, new_t, new_b)
+    return Z
 end
 
 """
@@ -41,10 +43,11 @@ in either networks are also present in the final network.
 function Base.union(X::T, Y::T) where {T<:UnipartiteNetwork}
     new_s = union(species(X), species(Y))
     new_a = spzeros(_interaction_type(X), length(new_s), length(new_s))
+    Z = UnipartiteNetwork(new_a, new_s)
     for common in union(interactions(X), interactions(Y))
-        new_a[common.from,common.to] = true
+        Z[common.from,common.to] = true
     end
-    return UnipartiteNetwork(new_a, new_s)
+    return Z
 end
 
 """
@@ -58,10 +61,11 @@ common.
 function Base.intersect(X::T, Y::T) where {T<:UnipartiteNetwork}
     new_s = intersect(species(X), species(Y))
     new_a = spzeros(_interaction_type(X), length(new_s), length(new_s))
+    Z = UnipartiteNetwork(new_a, new_s)
     for common in intersect(interactions(X), interactions(Y))
-        new_a[common.from,common.to] = true
+        Z[common.from,common.to] = true
     end
-    return UnipartiteNetwork(new_a, new_s)
+    return Z
 end
 
 """
