### 2020-11-30 6691d51
📖 update docstrings
