+++ src/betadiversity/operations.jl
@@ -7,12 +7,9 @@ in either networks are also present in the final network.
 function Base.union(X::T, Y::T) where {T<:BipartiteNetwork}
     new_t = union(species(X; dims=1), species(Y; dims=1))
     new_b = union(species(X; dims=2), species(Y; dims=2))
-    new_a = spzeros(_interaction_type(X), length(new_t), length(new_b))
-    Z = BipartiteNetwork(new_a, new_t, new_b)
-    for common in union(interactions(X), interactions(Y))
-        Z[common.from,common.to] = true
-    end
-    return Z
+    int = interactions(X) ∪ interactions(Y)
+    new_a = sparse([i.from for i in int], [i.to for i in int], true, length(new_t), length(new_b))
+    return BipartiteNetwork(new_a, new_t, new_b)
 end
 
 """
@@ -26,12 +23,9 @@ common.
 function Base.intersect(X::T, Y::T) where {T<:BipartiteNetwork}
     new_t = intersect(species(X; dims=1), species(Y; dims=1))
     new_b = intersect(species(X; dims=2), species(Y; dims=2))
-    new_a = spzeros(_interaction_type(X), length(new_t), length(new_b))
-    Z = BipartiteNetwork(new_a, new_t, new_b)
-    for common in intersect(interactions(X), interactions(Y))
-        Z[common.from,common.to] = true
-    end
-    return Z
+    int = interactions(X) ∩ interactions(Y)
+    new_a = sparse([i.from for i in int], [i.to for i in int], true, length(new_t), length(new_b))
+    return BipartiteNetwork(new_a, new_t, new_b)
 end
 
 """
@@ -42,12 +36,9 @@ in either networks are also present in the final network.
 """
 function Base.union(X::T, Y::T) where {T<:UnipartiteNetwork}
     new_s = union(species(X), species(Y))
-    new_a = spzeros(_interaction_type(X), length(new_s), length(new_s))
-    Z = UnipartiteNetwork(new_a, new_s)
-    for common in union(interactions(X), interactions(Y))
-        Z[common.from,common.to] = true
-    end
-    return Z
+    int = interactions(X) ∪ interactions(Y)
+    new_a = sparse([i.from for i in int], [i.to for i in int], true, length(new_s), length(new_s))
+    return UnipartiteNetwork(new_a, new_s)
 end
 
 """
@@ -60,12 +51,9 @@ common.
 """
 function Base.intersect(X::T, Y::T) where {T<:UnipartiteNetwork}
     new_s = intersect(species(X), species(Y))
-    new_a = spzeros(_interaction_type(X), length(new_s), length(new_s))
-    Z = UnipartiteNetwork(new_a, new_s)
-    for common in intersect(interactions(X), interactions(Y))
-        Z[common.from,common.to] = true
-    end
-    return Z
+    int = interactions(X) ∩ interactions(Y)
+    new_a = sparse([i.from for i in int], [i.to for i in int], true, length(new_s), length(new_s))
+    return UnipartiteNetwork(new_a, new_s)
 end
 
 """
