+++ src/community/centrality.jl
@@ -18,15 +18,17 @@ function centrality_katz(N::Union{UnipartiteNetwork, UnipartiteProbabilisticNetw
 end
 
 """
-    centrality_closeness(N::UnipartiteNetwork; nmax::Int64=100)
+    centrality_closeness(N::UnipartiteNetwork; nmax::Int64=20)
 
 The function calls `shortest_path` internally -- the `nmax` argument is the
 maximal path length that will be tried.
 
-> Bavelas, A., 1950. Communication Patterns in Task‐Oriented Groups. The Journal
-> of the Acoustical Society of America 22, 725–730. doi:10.1121/1.1906679
+#### References
+
+- Bavelas, A., 1950. Communication Patterns in Task‐Oriented Groups. The Journal
+  of the Acoustical Society of America 22, 725–730. doi:10.1121/1.1906679
 """
-function centrality_closeness(N::UnipartiteNetwork; nmax::Int64=100)
+function centrality_closeness(N::UnipartiteNetwork; nmax::Int64=20)
   d = shortest_path(N, nmax=nmax)
   n = richness(N)-1
   d[diagind(d)] .= 0
@@ -35,6 +37,24 @@ function centrality_closeness(N::UnipartiteNetwork; nmax::Int64=100)
   for i in eachindex(interm)
     interm[i] = interm[i] == Inf ? 0.0 : interm[i]
   end
+  return Dict(zip(species(N), interm))
+end
+
+"""
+    centrality_harmonic(N::UnipartiteNetwork; nmax::Int64=20)
+
+The function calls `shortest_path` internally -- the `nmax` argument is the
+maximal path length that will be tried.
+"""
+function centrality_harmonic(N::UnipartiteNetwork; nmax::Int64=20)
+  d = shortest_path(N, nmax=nmax)
+  n = richness(N)-1
+  d[diagind(d)] .= 0
+  interm = 1.0 ./ d
+  for i in eachindex(interm)
+    interm[i] = isinf(interm[i]) ? 0.0 : interm[i]
+  end
+  interm = sum(d; dims=2)./n
   return Dict(zip(species(N), interm))
 end
 
@@ -48,4 +68,19 @@ function centrality_degree(N::UnipartiteNetwork)
   d = degree(N)
   dm = maximum(values(d))
   return Dict([p.first=>p.second/dm for p in d])
+end
+
+
+"""
+    centrality_eigenvector(N::AbstractUnipartiteNetwork)
+
+Eigen centrality, corrected by the maximum degree (the most central species has
+a degree of 1).
+"""
+function centrality_eigenvector(N::AbstractUnipartiteNetwork)
+  evals, evecs = eigen(Array(N.edges))
+  emax, epos = findmax(real.(evals))
+  cvals = vec(real.(evecs[:,epos]))
+  cvals ./= maximum(cvals)
+  return Dict(zip(species(N), cvals))
 end
