+++ docs/src/interface/highlevel.md
@@ -42,4 +42,3 @@ nodiagonal!
 ```@docs
 Base.:!
 ```
-
