### 2021-01-27 84e4abb
start adding community detection functions
