+++ src/modularity/walktrap.jl
@@ -1,0 +1,5 @@
+"""
+
+Note that by contrast to the original publication, we assume that the network is
+directed, and that nodes do not necessarily have self-edges.
+"""
