+++ docs/src/properties/modularity.md
@@ -30,6 +30,7 @@ lp
 ```@docs
 brim
 salp
+leadingeigenvector
 ```
 
 ## Functional roles
