### 2021-02-19 f13754a
⚡ faster linear filter (and so null models)
