+++ src/rand/linfilter.jl
@@ -6,6 +6,11 @@ Given a network `N` compute the linear filter scores according to Stock et al.
 or missing interactions. Though this it returned as a probabilistic network,
 score do not necessary convey a probabilistic interpretation.
 
+The values of α give the *relative* weight of, in order, the measured
+interaction value, the out-degree of the species, the in-degree of the species,
+and of network connectance. The default parameterization is to have all four at
+equal importance.
+
 #### References
 
 Stock, M., Pahikkala, T., Airola, A., Waegeman, W., Baets, B.D., 2018. Algebraic
@@ -16,38 +21,19 @@ Stock, M., Poisot, T., Waegeman, W., Baets, B.D., 2017. Linear filtering reveals
 false negatives in species interaction data. Scientific Reports 7, 45908.
 https://doi.org/10.1038/srep45908
 """
-function linearfilter(N::T; α::Vector{Float64}=fill(0.25, 4)) where {T <: BinaryNetwork}
-  @assert length(α) == 4
-  @assert all(α .≥ 0.0)
-  α = α./sum(α) # This ensures that α sums to 1.0
-
-  # Get the size of the network
-  n = richness(N; dims=1)
-  m = richness(N; dims=2)
-
-  # Get the in and out degree, as well as the total number of interactions
-  k_out = degree(N; dims=1)
-  k_in = degree(N; dims=2)
-  A = links(N)
-
-  # Prepare a return object
-  return_type = T <: AbstractBipartiteNetwork ? BipartiteProbabilisticNetwork : UnipartiteProbabilisticNetwork
-  F = return_type(
-    zeros(Float64, size(N)),
-    EcologicalNetworks._species_objects(N)...
-  )
-
-  # Get probabilities
-  for i in 1:richness(N; dims=1)
-    s1 = species(N; dims=1)[i]
-    for j in 1:richness(N; dims=2)
-      s2 = species(N; dims=2)[j]
-      t_val = α[1]*N[i,j] + (α[2]/m)*k_out[s1] + (α[3]/n)*k_in[s2] + α[4]/(n*m)*A
-      F[s1,s2] = max(min(t_val, 1.0), 0.0)
-    end
-  end
-
-  return F
+function linearfilter(N::T; α::Vector{Float64}=fill(1.0, 4)) where {T <: BinaryNetwork}
+    @assert length(α) == 4
+    @assert all(α .≥ 0.0)
+    α = α ./ sum(α) # This ensures that α sums to 1.0
+    
+    # Get the new weights
+    W = α[1]A.edges .+ α[2]mean(A.edges; dims=2) .+ α[3]mean(A.edges; dims=1) .+ α[4]mean(A.edges)
+    
+    # Prepare a return object
+    return_type = T <: AbstractBipartiteNetwork ? BipartiteProbabilisticNetwork : UnipartiteProbabilisticNetwork
+    F = return_type(W, EcologicalNetworks._species_objects(N)...)
+    
+    return W
 end
 
 """
@@ -71,20 +57,20 @@ false negatives in species interaction data. Scientific Reports 7, 45908.
 https://doi.org/10.1038/srep45908
 """
 function linearfilterzoo(N::T; α::Vector{Float64}=fill(0.25, 4)) where {T <: BinaryNetwork}
-  F = linearfilter(N, α=α)  # depart from scores of the filter
-  Fzoo = F  # update the values to zero-one-out scores, reference is to keep
-            # the code readable
-
-  # Get the size of the network
-  n = richness(N; dims=1)
-  m = richness(N; dims=2)
-  # Get probabilities
-  for s1 in species(N; dims=1)
[…]
