+++ src/modularity/walktrap.jl
@@ -1,5 +1,0 @@
-"""
-
-Note that by contrast to the original publication, we will not assume that nodes
-have same edges - but we will assume that the network is undirected.
-"""
