+++ docs/src/properties/modularity.md
@@ -30,7 +30,6 @@ lp
 ```@docs
 brim
 salp
-leadingeigenvector
 ```
 
 ## Functional roles
