+++ Project.toml
@@ -23,9 +23,3 @@ Distributions = "0.23, 0.24"
 Requires = "1.0"
 StatsBase = "0.33"
 julia = "1.4"
-
-[extras]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[targets]
-test = ["Test"]
