+++ src/EcologicalNetworks.jl
@@ -10,7 +10,6 @@ using DelimitedFiles
 using LinearAlgebra
 using DataStructures
 using SparseArrays
-
 using Requires
 
 # Various utilities for probabilities
