### 2021-02-19 c644725
✨ rename truncated svd to rdpg
