### 2021-02-19 #2721 @KristinaRiemer
@tpoisot yes, I think at this point we should look for another reviewer. Hopefully @kellykochanski is doing alright!

If you have suggestions of reviewers with sufficient expertise that adhere to the [JOSS conflict of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) guidelines, that would be great! I will also check for reviewers from our JOSS list.
