### 2021-02-20 #2721 @tpoisot
@KristinaRiemer all the people I would think of from the top of my head have COI - but I can think about it some more.

That being said, my understanding is that @kellykochanski opened a number of issues on the repo as part of their review,

EcoJulia/EcologicalNetworks.jl#166 - contribution and authorship
EcoJulia/EcologicalNetworks.jl#167 - installation instructions
EcoJulia/EcologicalNetworks.jl#168 - software paper

and that @FrancisBanville addressed these comments with modifications. Could you confirm that the responses are satisfactory, and given that we've had another review with all comments resolved, can you make an editorial decision based on this?
