### 2021-02-25 68c2ff4
:racehorse: Calculate Q to save memory
