### 2021-03-05 8457bc6
:racehorse: Preprocessing with adjacent vertices
