### 2021-03-06 6d54eb7
🎨 replace the first loop
