### 2021-04-21 #2721 @KristinaRiemer
Okay, I finished up the review started by @kellykochanski and everything looks good! I especially like the diagram of the package workflow in the paper. Are you leaving the paper in the `joss-article-review1` branch?

Otherwise, will say this paper is accepted. We'll have a few more steps to go through, including a last review of the paper itself and getting a DOI.
