### 2021-04-21 #2721 @KristinaRiemer
Everything in the paper looks good except a couple of the references. Could you edit the Lightgraph reference so that the other contributors doesn't look like an author? And I think the DOI for the "Nestedness versus modularity in ecological networks: two sides of the same coin?" should be changed to https://doi.org/10.1111/j.1365-2656.2010.01688.x, so it doesn't show up as missing?
