+++ paper/paper.bib
@@ -43,7 +43,7 @@
 
 @software{BromOthe17,
   title = {Juliagraphs/{{Lightgraphs}}.{{Jl}}},
-  author = {Bromberger, Seth and Other Contributors},
+  author = {Bromberger, Seth},
   date = {2017-09-16},
   doi = {10.5281/ZENODO.889971},
   url = {https://zenodo.org/record/889971},
@@ -176,6 +176,7 @@
   pages = {811--817},
   publisher = {{[Wiley, British Ecological Society]}},
   issn = {0021-8790},
+  doi = {10.1111/j.1365-2656.2010.01688.x},
   abstract = {1. Understanding the structure of ecological networks is a crucial task for interpreting community and ecosystem responses to global change. 2. Despite the recent interest in this subject, almost all studies have focused exclusively on one specific network property. The question remains as to what extent different network properties are related and how understanding this relationship can advance our comprehension of the mechanisms behind these patterns. 3. Here, we analysed the relationship between nestedness and modularity, two frequently studied network properties, for a large data set of 95 ecological communities including both plant-animal mutualistic and host-parasite networks. 4. We found that the correlation between nestedness and modularity for a population of random matrices generated from the real communities decreases significantly in magnitude and sign with increasing connectance independent of the network type. At low connectivities, networks that are highly nested also tend to be highly modular; the reverse happens at high connectivities. 5. The above result is qualitatively robust when different null models are used to infer network structure, but, at a finer scale, quantitative differences exist. We observed an important interaction between the network structure pattern and the null model used to detect it. 6. A better understanding of the relationship between nestedness and modularity is important given their potential implications on the dynamics and stability of ecological communities.},
   eprint = {40835468},
   eprinttype = {jstor},
