### 2021-04-23 #2721 @KristinaRiemer
That looks great, thanks for doing that!

The next steps are to: 1) generate a new tagged release on GitHub and 2) archive the code (on Zenodo, figshare, institutional repository, etc.) to generate a DOI. If you can post the tag version and a link to the DOI here when you're done please.
