### 2021-04-26 #2721 @FrancisBanville
Hello @KristinaRiemer,
The tagged versions are [v0.4.3 for EcologicalNetworks.jl ](https://github.com/EcoJulia/EcologicalNetworks.jl/releases/tag/v0.4.3) and [v0.3.1 for Mangal.jl](https://github.com/EcoJulia/Mangal.jl/releases/tag/v0.3.1).
We archived our repositories and the code to reproduce our paper on Zenodo [at this link](https://zenodo.org/record/4716483#.YIb5mJBKiM9).

Please let me know if everything is okay.
Thank you!
