### 2021-04-26 #2721 @KristinaRiemer
Great, I'll use the EcologicalNetworks version here. Could you change the Zenodo title to match the software paper title?
